Separate logic for timer task execution and queue management
