+++ service/history/historyEngineInterfaces.go
@@ -74,6 +74,12 @@ type (
 		GetDomainID() string
 	}
 
+	queueTaskExecutor interface {
+		execute(taskInfo *taskInfo) error
+	}
+
+	// TODO: deprecate this interface in favor of the task interface
+	// defined in common/task package
 	taskExecutor interface {
 		process(taskInfo *taskInfo) (int, error)
 		complete(taskInfo *taskInfo)
