+++ service/history/timerQueueActiveProcessor.go
@@ -21,16 +21,11 @@
 package history
 
 import (
-	"fmt"
 	"time"
 
 	"github.com/pborman/uuid"
 
-	m "github.com/uber/cadence/.gen/go/matching"
-	workflow "github.com/uber/cadence/.gen/go/shared"
 	"github.com/uber/cadence/client/matching"
-	"github.com/uber/cadence/common"
-	"github.com/uber/cadence/common/backoff"
 	"github.com/uber/cadence/common/log"
 	"github.com/uber/cadence/common/log/tag"
 	"github.com/uber/cadence/common/metrics"
@@ -40,16 +35,13 @@ import (
 type (
 	timerQueueActiveProcessorImpl struct {
 		shard                   ShardContext
-		historyService          *historyEngineImpl
-		cache                   *historyCache
 		timerTaskFilter         taskFilter
 		now                     timeNow
 		logger                  log.Logger
 		metricsClient           metrics.Client
 		currentClusterName      string
-		matchingClient          matching.Client
+		taskExecutor            queueTaskExecutor
 		timerQueueProcessorBase *timerQueueProcessorBase
-		config                  *Config
 	}
 )
 
@@ -91,12 +83,9 @@ func newTimerQueueActiveProcessor(
 	timerGate := NewLocalTimerGate(shard.GetTimeSource())
 	processor := &timerQueueActiveProcessorImpl{
 		shard:              shard,
-		historyService:     historyService,
-		cache:              historyService.historyCache,
 		timerTaskFilter:    timerTaskFilter,
 		now:                timeNow,
 		logger:             logger,
-		matchingClient:     matchingClient,
 		metricsClient:      historyService.metricsClient,
 		currentClusterName: currentClusterName,
 		timerQueueProcessorBase: newTimerQueueProcessorBase(
@@ -108,9 +97,16 @@ func newTimerQueueActiveProcessor(
 			shard.GetConfig().TimerProcessorMaxPollRPS,
 			logger,
 		),
-		config: shard.GetConfig(),
 	}
 	processor.timerQueueProcessorBase.timerProcessor = processor
+	processor.taskExecutor = newTimerQueueActiveTaskExecutor(
+		shard,
+		historyService,
+		processor,
+		logger,
+		historyService.metricsClient,
+		shard.GetConfig(),
+	)
 	return processor
 }
 
@@ -177,13 +173,10 @@ func newTimerQueueFailoverProcessor(
 	timerGate := NewLocalTimerGate(shard.GetTimeSource())
 	processor := &timerQueueActiveProcessorImpl{
 		shard:              shard,
-		historyService:     historyService,
-		cache:              historyService.historyCache,
 		timerTaskFilter:    timerTaskFilter,
 		now:                timeNow,
 		logger:             logger,
 		metricsClient:      historyService.metricsClient,
-		matchingClient:     matchingClient,
 		currentClusterName: currentClusterName,
 		timerQueueProcessorBase: newTimerQueueProcessorBase(
 			metrics.TimerActiveQueueProcessorScope,
@@ -194,7 +187,6 @@ func newTimerQueueFailoverProcessor(
 			shard.GetConfig().TimerProcessorFailoverMaxPollRPS,
 			logger,
 		),
-		config: shard.GetConfig(),
 	}
 	processor.timerQueueProcessorBase.timerProcessor = processor
 	return updateShardAckLevel, processor
@@ -241,558 +233,566 @@ func (t *timerQueueActiveProcessorImpl) complete(
 func (t *timerQueueActiveProcessorImpl) process(
 	taskInfo *taskInfo,
 ) (int, error) {
-
-	timerTask, ok := taskInfo.task.(*persistence.TimerTaskInfo)
-	if !ok {
-		return metrics.TimerActiveQueueProcessorScope, errUnexpectedQueueTask
-	}
-
-	var err error
-	switch timerTask.TaskType {
-	case persistence.TaskTypeUserTimer:
-		if taskInfo.shouldProcessTask {
-			err = t.processUserTimerTimeout(timerTask)
-		}
-		return metrics.TimerActiveTaskUserTimerScope, err
-
-	case persistence.TaskTypeActivityTimeout:
-		if taskInfo.shouldProcessTask {
-			err = t.processActivityTimeout(timerTask)
-		}
-		return metrics.TimerActiveTaskActivityTimeoutScope, err
-
-	case persistence.TaskTypeDecisionTimeout:
-		if taskInfo.shouldProcessTask {
-			err = t.processDecisionTimeout(timerTask)
-		}
-		return metrics.TimerActiveTaskDecisionTimeoutScope, err
-
-	case persistence.TaskTypeWorkflowTimeout:
-		if taskInfo.shouldProcessTask {
-			err = t.processWorkflowTimeout(timerTask)
-		}
-		return metrics.TimerActiveTaskWorkflowTimeoutScope, err
-
-	case persistence.TaskTypeActivityRetryTimer:
-		if taskInfo.shouldProcessTask {
-			err = t.processActivityRetryTimer(timerTask)
-		}
-		return metrics.TimerActiveTaskActivityRetryTimerScope, err
-
-	case persistence.TaskTypeWorkflowBackoffTimer:
-		if taskInfo.shouldProcessTask {
-			err = t.processWorkflowBackoffTimer(timerTask)
-		}
-		return metrics.TimerActiveTaskWorkflowBackoffTimerScope, err
-
-	case persistence.TaskTypeDeleteHistoryEvent:
-		if taskInfo.shouldProcessTask {
-			err = t.timerQueueProcessorBase.processDeleteHistoryEvent(timerTask)
-		}
-		return metrics.TimerActiveTaskDeleteHistoryEventScope, err
-
-	default:
-		return metrics.TimerActiveQueueProcessorScope, errUnknownTimerTask
-	}
+	// TODO: task metricScope should be determined when creating taskInfo
+	metricScope := t.timerQueueProcessorBase.getTimerTaskMetricScope(taskInfo.task.GetTaskType(), true)
+	return metricScope, t.taskExecutor.execute(taskInfo)
 }
 
-func (t *timerQueueActiveProcessorImpl) processUserTimerTimeout(
-	task *persistence.TimerTaskInfo,
-) (retError error) {
+// func (t *timerQueueActiveProcessorImpl) process(
+// 	taskInfo *taskInfo,
+// ) (int, error) {
 
-	context, release, err := t.cache.getOrCreateWorkflowExecutionForBackground(
-		t.timerQueueProcessorBase.getDomainIDAndWorkflowExecution(task),
-	)
-	if err != nil {
-		return err
-	}
-	defer func() { release(retError) }()
+// 	timerTask, ok := taskInfo.task.(*persistence.TimerTaskInfo)
+// 	if !ok {
+// 		return metrics.TimerActiveQueueProcessorScope, errUnexpectedQueueTask
+// 	}
 
-	mutableState, err := loadMutableStateForTimerTask(context, task, t.metricsClient, t.logger)
-	if err != nil {
-		return err
-	}
-	if mutableState == nil || !mutableState.IsWorkflowExecutionRunning() {
-		return nil
-	}
+// 	var err error
+// 	switch timerTask.TaskType {
+// 	case persistence.TaskTypeUserTimer:
+// 		if taskInfo.shouldProcessTask {
+// 			err = t.processUserTimerTimeout(timerTask)
+// 		}
+// 		return metrics.TimerActiveTaskUserTimerScope, err
 
-	timerSequence := t.getTimerSequence(mutableState)
-	referenceTime := t.shard.GetTimeSource().Now()
-	timerFired := false
+// 	case persistence.TaskTypeActivityTimeout:
+// 		if taskInfo.shouldProcessTask {
+// 			err = t.processActivityTimeout(timerTask)
+// 		}
+// 		return metrics.TimerActiveTaskActivityTimeoutScope, err
 
-Loop:
-	for _, timerSequenceID := range timerSequence.loadAndSortUserTimers() {
-		timerInfo, ok := mutableState.GetUserTimerInfoByEventID(timerSequenceID.eventID)
-		if !ok {
-			errString := fmt.Sprintf("failed to find in user timer event ID: %v", timerSequenceID.eventID)
-			t.logger.Error(errString)
-			return &workflow.InternalServiceError{Message: errString}
-		}
+// 	case persistence.TaskTypeDecisionTimeout:
+// 		if taskInfo.shouldProcessTask {
+// 			err = t.processDecisionTimeout(timerTask)
+// 		}
+// 		return metrics.TimerActiveTaskDecisionTimeoutScope, err
 
-		if expired := timerSequence.isExpired(referenceTime, timerSequenceID); !expired {
-			// timer sequence IDs are sorted, once there is one timer
-			// sequence ID not expired, all after that wil not expired
-			break Loop
-		}
+// 	case persistence.TaskTypeWorkflowTimeout:
+// 		if taskInfo.shouldProcessTask {
+// 			err = t.processWorkflowTimeout(timerTask)
+// 		}
+// 		return metrics.TimerActiveTaskWorkflowTimeoutScope, err
 
-		if _, err := mutableState.AddTimerFiredEvent(timerInfo.TimerID); err != nil {
-			return err
-		}
-		timerFired = true
-	}
+// 	case persistence.TaskTypeActivityRetryTimer:
+// 		if taskInfo.shouldProcessTask {
+// 			err = t.processActivityRetryTimer(timerTask)
+// 		}
+// 		return metrics.TimerActiveTaskActivityRetryTimerScope, err
 
-	if !timerFired {
-		return nil
-	}
+// 	case persistence.TaskTypeWorkflowBackoffTimer:
+// 		if taskInfo.shouldProcessTask {
+// 			err = t.processWorkflowBackoffTimer(timerTask)
+// 		}
+// 		return metrics.TimerActiveTaskWorkflowBackoffTimerScope, err
 
-	return t.updateWorkflowExecution(context, mutableState, timerFired)
-}
+// 	case persistence.TaskTypeDeleteHistoryEvent:
+// 		if taskInfo.shouldProcessTask {
+// 			err = t.timerQueueProcessorBase.processDeleteHistoryEvent(timerTask)
+// 		}
+// 		return metrics.TimerActiveTaskDeleteHistoryEventScope, err
 
-func (t *timerQueueActiveProcessorImpl) processActivityTimeout(
-	task *persistence.TimerTaskInfo,
-) (retError error) {
+// 	default:
+// 		return metrics.TimerActiveQueueProcessorScope, errUnknownTimerTask
+// 	}
+// }
 
-	context, release, err := t.cache.getOrCreateWorkflowExecutionForBackground(
-		t.timerQueueProcessorBase.getDomainIDAndWorkflowExecution(task),
-	)
-	if err != nil {
-		return err
-	}
-	defer func() { release(retError) }()
+// func (t *timerQueueActiveProcessorImpl) processUserTimerTimeout(
+// 	task *persistence.TimerTaskInfo,
+// ) (retError error) {
 
-	mutableState, err := loadMutableStateForTimerTask(context, task, t.metricsClient, t.logger)
-	if err != nil {
-		return err
-	}
-	if mutableState == nil || !mutableState.IsWorkflowExecutionRunning() {
-		return nil
-	}
+// 	context, release, err := t.cache.getOrCreateWorkflowExecutionForBackground(
+// 		t.timerQueueProcessorBase.getDomainIDAndWorkflowExecution(task),
+// 	)
+// 	if err != nil {
+// 		return err
+// 	}
+// 	defer func() { release(retError) }()
 
-	timerSequence := t.getTimerSequence(mutableState)
-	referenceTime := t.shard.GetTimeSource().Now()
-	updateMutableState := false
-	scheduleDecision := false
+// 	mutableState, err := loadMutableStateForTimerTask(context, task, t.metricsClient, t.logger)
+// 	if err != nil {
+// 		return err
+// 	}
+// 	if mutableState == nil || !mutableState.IsWorkflowExecutionRunning() {
+// 		return nil
+// 	}
 
-	// need to clear activity heartbeat timer task mask for new activity timer task creation
-	// NOTE: LastHeartbeatTimeoutVisibility is for deduping heartbeat timer creation as it's possible
-	// one heartbeat task was persisted multiple times with different taskIDs due to the retry logic
-	// for updating workflow execution. In that case, only one new heartbeat timeout task should be
-	// created.
-	isHeartBeatTask := task.TimeoutType == int(workflow.TimeoutTypeHeartbeat)
-	activityInfo, ok := mutableState.GetActivityInfo(task.EventID)
-	if isHeartBeatTask && ok && activityInfo.LastHeartbeatTimeoutVisibility <= task.VisibilityTimestamp.Unix() {
-		activityInfo.TimerTaskStatus = activityInfo.TimerTaskStatus &^ timerTaskStatusCreatedHeartbeat
-		if err := mutableState.UpdateActivity(activityInfo); err != nil {
-			return err
-		}
-		updateMutableState = true
-	}
+// 	timerSequence := t.getTimerSequence(mutableState)
+// 	referenceTime := t.shard.GetTimeSource().Now()
+// 	timerFired := false
 
-Loop:
-	for _, timerSequenceID := range timerSequence.loadAndSortActivityTimers() {
-		activityInfo, ok := mutableState.GetActivityInfo(timerSequenceID.eventID)
-		if !ok || timerSequenceID.attempt < activityInfo.Attempt {
-			// handle 2 cases:
-			// 1. !ok
-			//  this case can happen since each activity can have 4 timers
-			//  and one of those 4 timers may have fired in this loop
-			// 2. timerSequenceID.attempt < activityInfo.Attempt
-			//  retry could update activity attempt, should not timeouts new attempt
-			continue Loop
-		}
+// Loop:
+// 	for _, timerSequenceID := range timerSequence.loadAndSortUserTimers() {
+// 		timerInfo, ok := mutableState.GetUserTimerInfoByEventID(timerSequenceID.eventID)
+// 		if !ok {
+// 			errString := fmt.Sprintf("failed to find in user timer event ID: %v", timerSequenceID.eventID)
+// 			t.logger.Error(errString)
+// 			return &workflow.InternalServiceError{Message: errString}
+// 		}
 
-		if expired := timerSequence.isExpired(referenceTime, timerSequenceID); !expired {
-			// timer sequence IDs are sorted, once there is one timer
-			// sequence ID not expired, all after that wil not expired
-			break Loop
-		}
+// 		if expired := timerSequence.isExpired(referenceTime, timerSequenceID); !expired {
+// 			// timer sequence IDs are sorted, once there is one timer
+// 			// sequence ID not expired, all after that wil not expired
+// 			break Loop
+// 		}
 
-		if timerSequenceID.timerType != timerTypeScheduleToStart {
-			// schedule to start timeout is not retriable
-			// customer should set larger schedule to start timeout if necessary
-			if ok, err := mutableState.RetryActivity(
-				activityInfo,
-				timerTypeToReason(timerSequenceID.timerType),
-				nil,
-			); err != nil {
-				return err
-			} else if ok {
-				updateMutableState = true
-				continue Loop
-			}
-		}
+// 		if _, err := mutableState.AddTimerFiredEvent(timerInfo.TimerID); err != nil {
+// 			return err
+// 		}
+// 		timerFired = true
+// 	}
 
-		t.emitTimeoutMetricScopeWithDomainTag(
-			mutableState.GetExecutionInfo().DomainID,
-			metrics.TimerActiveTaskActivityTimeoutScope,
-			timerSequenceID.timerType,
-		)
-		if _, err := mutableState.AddActivityTaskTimedOutEvent(
-			activityInfo.ScheduleID,
-			activityInfo.StartedID,
-			timerTypeToThrift(timerSequenceID.timerType),
-			activityInfo.Details,
-		); err != nil {
-			return err
-		}
-		updateMutableState = true
-		scheduleDecision = true
-	}
+// 	if !timerFired {
+// 		return nil
+// 	}
 
-	if !updateMutableState {
-		return nil
-	}
-	return t.updateWorkflowExecution(context, mutableState, scheduleDecision)
-}
+// 	return t.updateWorkflowExecution(context, mutableState, timerFired)
+// }
 
-func (t *timerQueueActiveProcessorImpl) processDecisionTimeout(
-	task *persistence.TimerTaskInfo,
-) (retError error) {
+// func (t *timerQueueActiveProcessorImpl) processActivityTimeout(
+// 	task *persistence.TimerTaskInfo,
+// ) (retError error) {
 
-	context, release, err := t.cache.getOrCreateWorkflowExecutionForBackground(
-		t.timerQueueProcessorBase.getDomainIDAndWorkflowExecution(task),
-	)
-	if err != nil {
-		return err
-	}
-	defer func() { release(retError) }()
+// 	context, release, err := t.cache.getOrCreateWorkflowExecutionForBackground(
+// 		t.timerQueueProcessorBase.getDomainIDAndWorkflowExecution(task),
+// 	)
+// 	if err != nil {
+// 		return err
+// 	}
+// 	defer func() { release(retError) }()
 
-	mutableState, err := loadMutableStateForTimerTask(context, task, t.metricsClient, t.logger)
-	if err != nil {
-		return err
-	}
-	if mutableState == nil || !mutableState.IsWorkflowExecutionRunning() {
-		return nil
-	}
+// 	mutableState, err := loadMutableStateForTimerTask(context, task, t.metricsClient, t.logger)
+// 	if err != nil {
+// 		return err
+// 	}
+// 	if mutableState == nil || !mutableState.IsWorkflowExecutionRunning() {
+// 		return nil
+// 	}
 
-	scheduleID := task.EventID
-	decision, ok := mutableState.GetDecisionInfo(scheduleID)
-	if !ok {
-		t.logger.Debug("Potentially duplicate task.", tag.TaskID(task.TaskID), tag.WorkflowScheduleID(scheduleID), tag.TaskType(persistence.TaskTypeDecisionTimeout))
-		return nil
-	}
-	ok, err = verifyTaskVersion(t.shard, t.logger, task.DomainID, decision.Version, task.Version, task)
-	if err != nil || !ok {
-		return err
-	}
+// 	timerSequence := t.getTimerSequence(mutableState)
+// 	referenceTime := t.shard.GetTimeSource().Now()
+// 	updateMutableState := false
+// 	scheduleDecision := false
 
-	if decision.Attempt != task.ScheduleAttempt {
-		return nil
-	}
+// 	// need to clear activity heartbeat timer task mask for new activity timer task creation
+// 	// NOTE: LastHeartbeatTimeoutVisibility is for deduping heartbeat timer creation as it's possible
+// 	// one heartbeat task was persisted multiple times with different taskIDs due to the retry logic
+// 	// for updating workflow execution. In that case, only one new heartbeat timeout task should be
+// 	// created.
+// 	isHeartBeatTask := task.TimeoutType == int(workflow.TimeoutTypeHeartbeat)
+// 	activityInfo, ok := mutableState.GetActivityInfo(task.EventID)
+// 	if isHeartBeatTask && ok && activityInfo.LastHeartbeatTimeoutVisibility <= task.VisibilityTimestamp.Unix() {
+// 		activityInfo.TimerTaskStatus = activityInfo.TimerTaskStatus &^ timerTaskStatusCreatedHeartbeat
+// 		if err := mutableState.UpdateActivity(activityInfo); err != nil {
+// 			return err
+// 		}
+// 		updateMutableState = true
+// 	}
 
-	scheduleDecision := false
-	switch timerTypeFromThrift(workflow.TimeoutType(task.TimeoutType)) {
-	case timerTypeStartToClose:
-		t.emitTimeoutMetricScopeWithDomainTag(
-			mutableState.GetExecutionInfo().DomainID,
-			metrics.TimerActiveTaskDecisionTimeoutScope,
-			timerTypeStartToClose,
-		)
-		if _, err := mutableState.AddDecisionTaskTimedOutEvent(
-			decision.ScheduleID,
-			decision.StartedID,
-		); err != nil {
-			return err
-		}
-		scheduleDecision = true
+// Loop:
+// 	for _, timerSequenceID := range timerSequence.loadAndSortActivityTimers() {
+// 		activityInfo, ok := mutableState.GetActivityInfo(timerSequenceID.eventID)
+// 		if !ok || timerSequenceID.attempt < activityInfo.Attempt {
+// 			// handle 2 cases:
+// 			// 1. !ok
+// 			//  this case can happen since each activity can have 4 timers
+// 			//  and one of those 4 timers may have fired in this loop
+// 			// 2. timerSequenceID.attempt < activityInfo.Attempt
+// 			//  retry could update activity attempt, should not timeouts new attempt
+// 			continue Loop
+// 		}
 
-	case timerTypeScheduleToStart:
-		if decision.StartedID != common.EmptyEventID {
-			// decision has already started
-			return nil
-		}
+// 		if expired := timerSequence.isExpired(referenceTime, timerSequenceID); !expired {
+// 			// timer sequence IDs are sorted, once there is one timer
+// 			// sequence ID not expired, all after that wil not expired
+// 			break Loop
+// 		}
 
-		t.emitTimeoutMetricScopeWithDomainTag(
-			mutableState.GetExecutionInfo().DomainID,
-			metrics.TimerActiveTaskDecisionTimeoutScope,
-			timerTypeScheduleToStart,
-		)
-		_, err := mutableState.AddDecisionTaskScheduleToStartTimeoutEvent(scheduleID)
-		if err != nil {
-			return err
-		}
-		scheduleDecision = true
-	}
+// 		if timerSequenceID.timerType != timerTypeScheduleToStart {
+// 			// schedule to start timeout is not retriable
+// 			// customer should set larger schedule to start timeout if necessary
+// 			if ok, err := mutableState.RetryActivity(
+// 				activityInfo,
+// 				timerTypeToReason(timerSequenceID.timerType),
+// 				nil,
+// 			); err != nil {
+// 				return err
+// 			} else if ok {
+// 				updateMutableState = true
+// 				continue Loop
+// 			}
+// 		}
 
-	return t.updateWorkflowExecution(context, mutableState, scheduleDecision)
-}
+// 		t.emitTimeoutMetricScopeWithDomainTag(
+// 			mutableState.GetExecutionInfo().DomainID,
+// 			metrics.TimerActiveTaskActivityTimeoutScope,
+// 			timerSequenceID.timerType,
+// 		)
+// 		if _, err := mutableState.AddActivityTaskTimedOutEvent(
+// 			activityInfo.ScheduleID,
+// 			activityInfo.StartedID,
+// 			timerTypeToThrift(timerSequenceID.timerType),
+// 			activityInfo.Details,
+// 		); err != nil {
+// 			return err
+// 		}
+// 		updateMutableState = true
+// 		scheduleDecision = true
+// 	}
 
-func (t *timerQueueActiveProcessorImpl) processWorkflowBackoffTimer(
-	task *persistence.TimerTaskInfo,
-) (retError error) {
+// 	if !updateMutableState {
+// 		return nil
+// 	}
+// 	return t.updateWorkflowExecution(context, mutableState, scheduleDecision)
+// }
 
-	context, release, err := t.cache.getOrCreateWorkflowExecutionForBackground(
-		t.timerQueueProcessorBase.getDomainIDAndWorkflowExecution(task),
-	)
-	if err != nil {
-		return err
-	}
-	defer func() { release(retError) }()
+// func (t *timerQueueActiveProcessorImpl) processDecisionTimeout(
+// 	task *persistence.TimerTaskInfo,
+// ) (retError error) {
 
-	mutableState, err := loadMutableStateForTimerTask(context, task, t.metricsClient, t.logger)
-	if err != nil {
-		return err
-	}
-	if mutableState == nil || !mutableState.IsWorkflowExecutionRunning() {
-		return nil
-	}
+// 	context, release, err := t.cache.getOrCreateWorkflowExecutionForBackground(
+// 		t.timerQueueProcessorBase.getDomainIDAndWorkflowExecution(task),
+// 	)
+// 	if err != nil {
+// 		return err
+// 	}
+// 	defer func() { release(retError) }()
 
-	if task.TimeoutType == persistence.WorkflowBackoffTimeoutTypeRetry {
-		t.metricsClient.IncCounter(metrics.TimerActiveTaskWorkflowBackoffTimerScope, metrics.WorkflowRetryBackoffTimerCount)
-	} else {
-		t.metricsClient.IncCounter(metrics.TimerActiveTaskWorkflowBackoffTimerScope, metrics.WorkflowCronBackoffTimerCount)
-	}
+// 	mutableState, err := loadMutableStateForTimerTask(context, task, t.metricsClient, t.logger)
+// 	if err != nil {
+// 		return err
+// 	}
+// 	if mutableState == nil || !mutableState.IsWorkflowExecutionRunning() {
+// 		return nil
+// 	}
 
-	if mutableState.HasProcessedOrPendingDecision() {
-		// already has decision task
-		return nil
-	}
+// 	scheduleID := task.EventID
+// 	decision, ok := mutableState.GetDecisionInfo(scheduleID)
+// 	if !ok {
+// 		t.logger.Debug("Potentially duplicate task.", tag.TaskID(task.TaskID), tag.WorkflowScheduleID(scheduleID), tag.TaskType(persistence.TaskTypeDecisionTimeout))
+// 		return nil
+// 	}
+// 	ok, err = verifyTaskVersion(t.shard, t.logger, task.DomainID, decision.Version, task.Version, task)
+// 	if err != nil || !ok {
+// 		return err
+// 	}
 
-	// schedule first decision task
-	return t.updateWorkflowExecution(context, mutableState, true)
-}
+// 	if decision.Attempt != task.ScheduleAttempt {
+// 		return nil
+// 	}
 
-func (t *timerQueueActiveProcessorImpl) processActivityRetryTimer(
-	task *persistence.TimerTaskInfo,
-) (retError error) {
+// 	scheduleDecision := false
+// 	switch timerTypeFromThrift(workflow.TimeoutType(task.TimeoutType)) {
+// 	case timerTypeStartToClose:
+// 		t.emitTimeoutMetricScopeWithDomainTag(
+// 			mutableState.GetExecutionInfo().DomainID,
+// 			metrics.TimerActiveTaskDecisionTimeoutScope,
+// 			timerTypeStartToClose,
+// 		)
+// 		if _, err := mutableState.AddDecisionTaskTimedOutEvent(
+// 			decision.ScheduleID,
+// 			decision.StartedID,
+// 		); err != nil {
+// 			return err
+// 		}
+// 		scheduleDecision = true
 
-	context, release, err := t.cache.getOrCreateWorkflowExecutionForBackground(
-		t.timerQueueProcessorBase.getDomainIDAndWorkflowExecution(task),
-	)
-	if err != nil {
-		return err
-	}
-	defer func() { release(retError) }()
+// 	case timerTypeScheduleToStart:
+// 		if decision.StartedID != common.EmptyEventID {
+// 			// decision has already started
+// 			return nil
+// 		}
 
-	mutableState, err := loadMutableStateForTimerTask(context, task, t.metricsClient, t.logger)
-	if err != nil {
-		return err
-	}
-	if mutableState == nil || !mutableState.IsWorkflowExecutionRunning() {
-		return nil
-	}
+// 		t.emitTimeoutMetricScopeWithDomainTag(
+// 			mutableState.GetExecutionInfo().DomainID,
+// 			metrics.TimerActiveTaskDecisionTimeoutScope,
+// 			timerTypeScheduleToStart,
+// 		)
+// 		_, err := mutableState.AddDecisionTaskScheduleToStartTimeoutEvent(scheduleID)
+// 		if err != nil {
+// 			return err
+// 		}
+// 		scheduleDecision = true
+// 	}
 
-	// generate activity task
-	scheduledID := task.EventID
-	activityInfo, ok := mutableState.GetActivityInfo(scheduledID)
-	if !ok || task.ScheduleAttempt < int64(activityInfo.Attempt) || activityInfo.StartedID != common.EmptyEventID {
-		if ok {
-			t.logger.Info("Duplicate activity retry timer task",
-				tag.WorkflowID(mutableState.GetExecutionInfo().WorkflowID),
-				tag.WorkflowRunID(mutableState.GetExecutionInfo().RunID),
-				tag.WorkflowDomainID(mutableState.GetExecutionInfo().DomainID),
-				tag.WorkflowScheduleID(activityInfo.ScheduleID),
-				tag.Attempt(activityInfo.Attempt),
-				tag.FailoverVersion(activityInfo.Version),
-				tag.TimerTaskStatus(activityInfo.TimerTaskStatus),
-				tag.ScheduleAttempt(task.ScheduleAttempt))
-		}
-		return nil
-	}
-	ok, err = verifyTaskVersion(t.shard, t.logger, task.DomainID, activityInfo.Version, task.Version, task)
-	if err != nil || !ok {
-		return err
-	}
+// 	return t.updateWorkflowExecution(context, mutableState, scheduleDecision)
+// }
 
-	domainID := task.DomainID
-	targetDomainID := domainID
-	if activityInfo.DomainID != "" {
-		targetDomainID = activityInfo.DomainID
-	} else {
-		// TODO remove this block after Mar, 1th, 2020
-		//  previously, DomainID in activity info is not used, so need to get
-		//  schedule event from DB checking whether activity to be scheduled
-		//  belongs to this domain
-		scheduledEvent, err := mutableState.GetActivityScheduledEvent(scheduledID)
-		if err != nil {
-			return err
-		}
-		if scheduledEvent.ActivityTaskScheduledEventAttributes.Domain != nil {
-			domainEntry, err := t.shard.GetDomainCache().GetDomain(scheduledEvent.ActivityTaskScheduledEventAttributes.GetDomain())
-			if err != nil {
-				return &workflow.InternalServiceError{Message: "unable to re-schedule activity across domain."}
-			}
-			targetDomainID = domainEntry.GetInfo().ID
-		}
-	}
+// func (t *timerQueueActiveProcessorImpl) processWorkflowBackoffTimer(
+// 	task *persistence.TimerTaskInfo,
+// ) (retError error) {
 
-	execution := workflow.WorkflowExecution{
-		WorkflowId: common.StringPtr(task.WorkflowID),
-		RunId:      common.StringPtr(task.RunID)}
-	taskList := &workflow.TaskList{
-		Name: common.StringPtr(activityInfo.TaskList),
-	}
-	scheduleToStartTimeout := activityInfo.ScheduleToStartTimeout
+// 	context, release, err := t.cache.getOrCreateWorkflowExecutionForBackground(
+// 		t.timerQueueProcessorBase.getDomainIDAndWorkflowExecution(task),
+// 	)
+// 	if err != nil {
+// 		return err
+// 	}
+// 	defer func() { release(retError) }()
 
-	release(nil) // release earlier as we don't need the lock anymore
+// 	mutableState, err := loadMutableStateForTimerTask(context, task, t.metricsClient, t.logger)
+// 	if err != nil {
+// 		return err
+// 	}
+// 	if mutableState == nil || !mutableState.IsWorkflowExecutionRunning() {
+// 		return nil
+// 	}
 
-	return t.matchingClient.AddActivityTask(nil, &m.AddActivityTaskRequest{
-		DomainUUID:                    common.StringPtr(targetDomainID),
-		SourceDomainUUID:              common.StringPtr(domainID),
-		Execution:                     &execution,
-		TaskList:                      taskList,
-		ScheduleId:                    common.Int64Ptr(scheduledID),
-		ScheduleToStartTimeoutSeconds: common.Int32Ptr(scheduleToStartTimeout),
-	})
-}
+// 	if task.TimeoutType == persistence.WorkflowBackoffTimeoutTypeRetry {
+// 		t.metricsClient.IncCounter(metrics.TimerActiveTaskWorkflowBackoffTimerScope, metrics.WorkflowRetryBackoffTimerCount)
+// 	} else {
+// 		t.metricsClient.IncCounter(metrics.TimerActiveTaskWorkflowBackoffTimerScope, metrics.WorkflowCronBackoffTimerCount)
+// 	}
 
-func (t *timerQueueActiveProcessorImpl) processWorkflowTimeout(
-	task *persistence.TimerTaskInfo,
-) (retError error) {
+// 	if mutableState.HasProcessedOrPendingDecision() {
+// 		// already has decision task
+// 		return nil
+// 	}
 
-	context, release, err := t.cache.getOrCreateWorkflowExecutionForBackground(
-		t.timerQueueProcessorBase.getDomainIDAndWorkflowExecution(task),
-	)
-	if err != nil {
-		return err
-	}
-	defer func() { release(retError) }()
+// 	// schedule first decision task
+// 	return t.updateWorkflowExecution(context, mutableState, true)
+// }
 
-	mutableState, err := loadMutableStateForTimerTask(context, task, t.metricsClient, t.logger)
-	if err != nil {
-		return err
-	}
-	if mutableState == nil || !mutableState.IsWorkflowExecutionRunning() {
-		return nil
-	}
+// func (t *timerQueueActiveProcessorImpl) processActivityRetryTimer(
+// 	task *persistence.TimerTaskInfo,
+// ) (retError error) {
 
-	startVersion, err := mutableState.GetStartVersion()
-	if err != nil {
-		return err
-	}
-	ok, err := verifyTaskVersion(t.shard, t.logger, task.DomainID, startVersion, task.Version, task)
-	if err != nil || !ok {
-		return err
-	}
+// 	context, release, err := t.cache.getOrCreateWorkflowExecutionForBackground(
+// 		t.timerQueueProcessorBase.getDomainIDAndWorkflowExecution(task),
+// 	)
+// 	if err != nil {
+// 		return err
+// 	}
+// 	defer func() { release(retError) }()
 
-	eventBatchFirstEventID := mutableState.GetNextEventID()
+// 	mutableState, err := loadMutableStateForTimerTask(context, task, t.metricsClient, t.logger)
+// 	if err != nil {
+// 		return err
+// 	}
+// 	if mutableState == nil || !mutableState.IsWorkflowExecutionRunning() {
+// 		return nil
+// 	}
 
-	timeoutReason := timerTypeToReason(timerTypeStartToClose)
-	backoffInterval := mutableState.GetRetryBackoffDuration(timeoutReason)
-	continueAsNewInitiator := workflow.ContinueAsNewInitiatorRetryPolicy
-	if backoffInterval == backoff.NoBackoff {
-		// check if a cron backoff is needed
-		backoffInterval, err = mutableState.GetCronBackoffDuration()
-		if err != nil {
-			return err
-		}
-		continueAsNewInitiator = workflow.ContinueAsNewInitiatorCronSchedule
-	}
-	if backoffInterval == backoff.NoBackoff {
-		if err := timeoutWorkflow(mutableState, eventBatchFirstEventID); err != nil {
-			return err
-		}
+// 	// generate activity task
+// 	scheduledID := task.EventID
+// 	activityInfo, ok := mutableState.GetActivityInfo(scheduledID)
+// 	if !ok || task.ScheduleAttempt < int64(activityInfo.Attempt) || activityInfo.StartedID != common.EmptyEventID {
+// 		if ok {
+// 			t.logger.Info("Duplicate activity retry timer task",
+// 				tag.WorkflowID(mutableState.GetExecutionInfo().WorkflowID),
+// 				tag.WorkflowRunID(mutableState.GetExecutionInfo().RunID),
+// 				tag.WorkflowDomainID(mutableState.GetExecutionInfo().DomainID),
+// 				tag.WorkflowScheduleID(activityInfo.ScheduleID),
+// 				tag.Attempt(activityInfo.Attempt),
+// 				tag.FailoverVersion(activityInfo.Version),
+// 				tag.TimerTaskStatus(activityInfo.TimerTaskStatus),
+// 				tag.ScheduleAttempt(task.ScheduleAttempt))
+// 		}
+// 		return nil
+// 	}
+// 	ok, err = verifyTaskVersion(t.shard, t.logger, task.DomainID, activityInfo.Version, task.Version, task)
+// 	if err != nil || !ok {
+// 		return err
+// 	}
 
-		// We apply the update to execution using optimistic concurrency.  If it fails due to a conflict than reload
-		// the history and try the operation again.
-		return t.updateWorkflowExecution(context, mutableState, false)
-	}
+// 	domainID := task.DomainID
+// 	targetDomainID := domainID
+// 	if activityInfo.DomainID != "" {
+// 		targetDomainID = activityInfo.DomainID
+// 	} else {
+// 		// TODO remove this block after Mar, 1th, 2020
+// 		//  previously, DomainID in activity info is not used, so need to get
+// 		//  schedule event from DB checking whether activity to be scheduled
+// 		//  belongs to this domain
+// 		scheduledEvent, err := mutableState.GetActivityScheduledEvent(scheduledID)
+// 		if err != nil {
+// 			return err
+// 		}
+// 		if scheduledEvent.ActivityTaskScheduledEventAttributes.Domain != nil {
+// 			domainEntry, err := t.shard.GetDomainCache().GetDomain(scheduledEvent.ActivityTaskScheduledEventAttributes.GetDomain())
+// 			if err != nil {
+// 				return &workflow.InternalServiceError{Message: "unable to re-schedule activity across domain."}
+// 			}
+// 			targetDomainID = domainEntry.GetInfo().ID
+// 		}
+// 	}
 
-	// workflow timeout, but a retry or cron is needed, so we do continue as new to retry or cron
-	startEvent, err := mutableState.GetStartEvent()
-	if err != nil {
-		return err
-	}
+// 	execution := workflow.WorkflowExecution{
+// 		WorkflowId: common.StringPtr(task.WorkflowID),
+// 		RunId:      common.StringPtr(task.RunID)}
+// 	taskList := &workflow.TaskList{
+// 		Name: common.StringPtr(activityInfo.TaskList),
+// 	}
+// 	scheduleToStartTimeout := activityInfo.ScheduleToStartTimeout
 
-	startAttributes := startEvent.WorkflowExecutionStartedEventAttributes
-	continueAsnewAttributes := &workflow.ContinueAsNewWorkflowExecutionDecisionAttributes{
-		WorkflowType:                        startAttributes.WorkflowType,
-		TaskList:                            startAttributes.TaskList,
-		Input:                               startAttributes.Input,
-		ExecutionStartToCloseTimeoutSeconds: startAttributes.ExecutionStartToCloseTimeoutSeconds,
-		TaskStartToCloseTimeoutSeconds:      startAttributes.TaskStartToCloseTimeoutSeconds,
-		BackoffStartIntervalInSeconds:       common.Int32Ptr(int32(backoffInterval.Seconds())),
-		RetryPolicy:                         startAttributes.RetryPolicy,
-		Initiator:                           continueAsNewInitiator.Ptr(),
-		FailureReason:                       common.StringPtr(timeoutReason),
-		CronSchedule:                        common.StringPtr(mutableState.GetExecutionInfo().CronSchedule),
-		Header:                              startAttributes.Header,
-		Memo:                                startAttributes.Memo,
-		SearchAttributes:                    startAttributes.SearchAttributes,
-	}
-	newMutableState, err := retryWorkflow(
-		mutableState,
-		eventBatchFirstEventID,
-		startAttributes.GetParentWorkflowDomain(),
-		continueAsnewAttributes,
-	)
-	if err != nil {
-		return err
-	}
+// 	release(nil) // release earlier as we don't need the lock anymore
 
-	newExecutionInfo := newMutableState.GetExecutionInfo()
-	return context.updateWorkflowExecutionWithNewAsActive(
-		t.shard.GetTimeSource().Now(),
-		newWorkflowExecutionContext(
-			newExecutionInfo.DomainID,
-			workflow.WorkflowExecution{
-				WorkflowId: common.StringPtr(newExecutionInfo.WorkflowID),
-				RunId:      common.StringPtr(newExecutionInfo.RunID),
-			},
-			t.shard,
-			t.shard.GetExecutionManager(),
-			t.logger,
-		),
-		newMutableState,
-	)
-}
+// 	return t.matchingClient.AddActivityTask(nil, &m.AddActivityTaskRequest{
+// 		DomainUUID:                    common.StringPtr(targetDomainID),
+// 		SourceDomainUUID:              common.StringPtr(domainID),
+// 		Execution:                     &execution,
+// 		TaskList:                      taskList,
+// 		ScheduleId:                    common.Int64Ptr(scheduledID),
+// 		ScheduleToStartTimeoutSeconds: common.Int32Ptr(scheduleToStartTimeout),
+// 	})
+// }
 
-func (t *timerQueueActiveProcessorImpl) getTimerSequence(
-	mutableState mutableState,
-) timerSequence {
+// func (t *timerQueueActiveProcessorImpl) processWorkflowTimeout(
+// 	task *persistence.TimerTaskInfo,
+// ) (retError error) {
 
-	timeSource := t.shard.GetTimeSource()
-	return newTimerSequence(timeSource, mutableState)
-}
+// 	context, release, err := t.cache.getOrCreateWorkflowExecutionForBackground(
+// 		t.timerQueueProcessorBase.getDomainIDAndWorkflowExecution(task),
+// 	)
+// 	if err != nil {
+// 		return err
+// 	}
+// 	defer func() { release(retError) }()
 
-func (t *timerQueueActiveProcessorImpl) updateWorkflowExecution(
-	context workflowExecutionContext,
-	mutableState mutableState,
-	scheduleNewDecision bool,
-) error {
+// 	mutableState, err := loadMutableStateForTimerTask(context, task, t.metricsClient, t.logger)
+// 	if err != nil {
+// 		return err
+// 	}
+// 	if mutableState == nil || !mutableState.IsWorkflowExecutionRunning() {
+// 		return nil
+// 	}
 
-	var err error
-	if scheduleNewDecision {
-		// Schedule a new decision.
-		err = scheduleDecision(mutableState)
-		if err != nil {
-			return err
-		}
-	}
+// 	startVersion, err := mutableState.GetStartVersion()
+// 	if err != nil {
+// 		return err
+// 	}
+// 	ok, err := verifyTaskVersion(t.shard, t.logger, task.DomainID, startVersion, task.Version, task)
+// 	if err != nil || !ok {
+// 		return err
+// 	}
 
-	now := t.shard.GetTimeSource().Now()
-	err = context.updateWorkflowExecutionAsActive(now)
-	if err != nil {
-		if isShardOwnershiptLostError(err) {
-			// Shard is stolen.  Stop timer processing to reduce duplicates
-			t.timerQueueProcessorBase.Stop()
-			return err
-		}
-		return err
-	}
+// 	eventBatchFirstEventID := mutableState.GetNextEventID()
 
-	return nil
-}
+// 	timeoutReason := timerTypeToReason(timerTypeStartToClose)
+// 	backoffInterval := mutableState.GetRetryBackoffDuration(timeoutReason)
+// 	continueAsNewInitiator := workflow.ContinueAsNewInitiatorRetryPolicy
+// 	if backoffInterval == backoff.NoBackoff {
+// 		// check if a cron backoff is needed
+// 		backoffInterval, err = mutableState.GetCronBackoffDuration()
+// 		if err != nil {
+// 			return err
+// 		}
+// 		continueAsNewInitiator = workflow.ContinueAsNewInitiatorCronSchedule
+// 	}
+// 	if backoffInterval == backoff.NoBackoff {
+// 		if err := timeoutWorkflow(mutableState, eventBatchFirstEventID); err != nil {
+// 			return err
+// 		}
 
-func (t *timerQueueActiveProcessorImpl) emitTimeoutMetricScopeWithDomainTag(
-	domainID string,
-	scope int,
-	timerType timerType,
-) {
+// 		// We apply the update to execution using optimistic concurrency.  If it fails due to a conflict than reload
+// 		// the history and try the operation again.
+// 		return t.updateWorkflowExecution(context, mutableState, false)
+// 	}
 
-	domainEntry, err := t.shard.GetDomainCache().GetDomainByID(domainID)
-	if err != nil {
-		return
-	}
-	metricsScope := t.metricsClient.Scope(scope).Tagged(metrics.DomainTag(domainEntry.GetInfo().Name))
-	switch timerType {
-	case timerTypeScheduleToStart:
-		metricsScope.IncCounter(metrics.ScheduleToStartTimeoutCounter)
-	case timerTypeScheduleToClose:
-		metricsScope.IncCounter(metrics.ScheduleToCloseTimeoutCounter)
-	case timerTypeStartToClose:
-		metricsScope.IncCounter(metrics.StartToCloseTimeoutCounter)
-	case timerTypeHeartbeat:
-		metricsScope.IncCounter(metrics.HeartbeatTimeoutCounter)
-	}
-}
+// 	// workflow timeout, but a retry or cron is needed, so we do continue as new to retry or cron
+// 	startEvent, err := mutableState.GetStartEvent()
+// 	if err != nil {
+// 		return err
+// 	}
+
+// 	startAttributes := startEvent.WorkflowExecutionStartedEventAttributes
+// 	continueAsnewAttributes := &workflow.ContinueAsNewWorkflowExecutionDecisionAttributes{
+// 		WorkflowType:                        startAttributes.WorkflowType,
+// 		TaskList:                            startAttributes.TaskList,
+// 		Input:                               startAttributes.Input,
+// 		ExecutionStartToCloseTimeoutSeconds: startAttributes.ExecutionStartToCloseTimeoutSeconds,
+// 		TaskStartToCloseTimeoutSeconds:      startAttributes.TaskStartToCloseTimeoutSeconds,
+// 		BackoffStartIntervalInSeconds:       common.Int32Ptr(int32(backoffInterval.Seconds())),
+// 		RetryPolicy:                         startAttributes.RetryPolicy,
+// 		Initiator:                           continueAsNewInitiator.Ptr(),
+// 		FailureReason:                       common.StringPtr(timeoutReason),
+// 		CronSchedule:                        common.StringPtr(mutableState.GetExecutionInfo().CronSchedule),
+// 		Header:                              startAttributes.Header,
+// 		Memo:                                startAttributes.Memo,
+// 		SearchAttributes:                    startAttributes.SearchAttributes,
+// 	}
+// 	newMutableState, err := retryWorkflow(
+// 		mutableState,
+// 		eventBatchFirstEventID,
+// 		startAttributes.GetParentWorkflowDomain(),
+// 		continueAsnewAttributes,
+// 	)
+// 	if err != nil {
+// 		return err
+// 	}
+
+// 	newExecutionInfo := newMutableState.GetExecutionInfo()
+// 	return context.updateWorkflowExecutionWithNewAsActive(
+// 		t.shard.GetTimeSource().Now(),
+// 		newWorkflowExecutionContext(
+// 			newExecutionInfo.DomainID,
+// 			workflow.WorkflowExecution{
+// 				WorkflowId: common.StringPtr(newExecutionInfo.WorkflowID),
+// 				RunId:      common.StringPtr(newExecutionInfo.RunID),
+// 			},
+// 			t.shard,
+// 			t.shard.GetExecutionManager(),
+// 			t.logger,
+// 		),
+// 		newMutableState,
+// 	)
+// }
+
+// func (t *timerQueueActiveProcessorImpl) getTimerSequence(
+// 	mutableState mutableState,
+// ) timerSequence {
+
+// 	timeSource := t.shard.GetTimeSource()
+// 	return newTimerSequence(timeSource, mutableState)
+// }
+
+// func (t *timerQueueActiveProcessorImpl) updateWorkflowExecution(
+// 	context workflowExecutionContext,
+// 	mutableState mutableState,
+// 	scheduleNewDecision bool,
+// ) error {
+
+// 	var err error
+// 	if scheduleNewDecision {
+// 		// Schedule a new decision.
+// 		err = scheduleDecision(mutableState)
+// 		if err != nil {
+// 			return err
+// 		}
+// 	}
+
+// 	now := t.shard.GetTimeSource().Now()
+// 	err = context.updateWorkflowExecutionAsActive(now)
+// 	if err != nil {
+// 		if isShardOwnershiptLostError(err) {
+// 			// Shard is stolen.  Stop timer processing to reduce duplicates
+// 			t.timerQueueProcessorBase.Stop()
+// 			return err
+// 		}
+// 		return err
+// 	}
+
+// 	return nil
+// }
+
+// func (t *timerQueueActiveProcessorImpl) emitTimeoutMetricScopeWithDomainTag(
+// 	domainID string,
+// 	scope int,
+// 	timerType timerType,
+// ) {
+
+// 	domainEntry, err := t.shard.GetDomainCache().GetDomainByID(domainID)
+// 	if err != nil {
+// 		return
+// 	}
+// 	metricsScope := t.metricsClient.Scope(scope).Tagged(metrics.DomainTag(domainEntry.GetInfo().Name))
+// 	switch timerType {
+// 	case timerTypeScheduleToStart:
+// 		metricsScope.IncCounter(metrics.ScheduleToStartTimeoutCounter)
+// 	case timerTypeScheduleToClose:
+// 		metricsScope.IncCounter(metrics.ScheduleToCloseTimeoutCounter)
+// 	case timerTypeStartToClose:
+// 		metricsScope.IncCounter(metrics.StartToCloseTimeoutCounter)
+// 	case timerTypeHeartbeat:
+// 		metricsScope.IncCounter(metrics.HeartbeatTimeoutCounter)
+// 	}
+// }
