+++ service/history/timerQueueActiveTaskExecutor_test.go
@@ -1,4 +1,4 @@
-// Copyright (c) 2019 Uber Technologies, Inc.
+// Copyright (c) 2020 Uber Technologies, Inc.
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -47,7 +47,7 @@ import (
 )
 
 type (
-	timerQueueActiveProcessorSuite struct {
+	timerQueueActiveTaskExecutorSuite struct {
 		suite.Suite
 		*require.Assertions
 
@@ -64,26 +64,26 @@ type (
 		mockExecutionMgr  *mocks.ExecutionManager
 		mockHistoryV2Mgr  *mocks.HistoryV2Manager
 
-		logger                    log.Logger
-		domainID                  string
-		domainEntry               *cache.DomainCacheEntry
-		version                   int64
-		now                       time.Time
-		timeSource                *clock.EventTimeSource
-		timerQueueActiveProcessor *timerQueueActiveProcessorImpl
+		logger                       log.Logger
+		domainID                     string
+		domainEntry                  *cache.DomainCacheEntry
+		version                      int64
+		now                          time.Time
+		timeSource                   *clock.EventTimeSource
+		timerQueueActiveTaskExecutor *timerQueueActiveTaskExecutor
 	}
 )
 
-func TestTimerQueueActiveProcessorSuite(t *testing.T) {
-	s := new(timerQueueActiveProcessorSuite)
+func TestTimerQueueActiveTaskExecutorSuite(t *testing.T) {
+	s := new(timerQueueActiveTaskExecutorSuite)
 	suite.Run(t, s)
 }
 
-func (s *timerQueueActiveProcessorSuite) SetupSuite() {
+func (s *timerQueueActiveTaskExecutorSuite) SetupSuite() {
 
 }
 
-func (s *timerQueueActiveProcessorSuite) SetupTest() {
+func (s *timerQueueActiveTaskExecutorSuite) SetupTest() {
 	s.Assertions = require.New(s.T())
 
 	s.domainID = testDomainID
@@ -100,13 +100,14 @@ func (s *timerQueueActiveProcessorSuite) SetupTest() {
 	s.mockReplicationProcessor.EXPECT().notifyNewTask().AnyTimes()
 	s.mockTimerProcessor.EXPECT().NotifyNewTimers(gomock.Any(), gomock.Any()).AnyTimes()
 
+	config := NewDynamicConfigForTest()
 	s.mockShard = newTestShardContext(
 		s.controller,
 		&persistence.ShardInfo{
 			RangeID:          1,
 			TransferAckLevel: 0,
 		},
-		NewDynamicConfigForTest(),
+		config,
 	)
 	s.mockShard.eventsCache = newEventsCache(s.mockShard)
 	s.mockShard.resource.TimeSource = s.timeSource
@@ -143,21 +144,28 @@ func (s *timerQueueActiveProcessorSuite) SetupTest() {
 	s.mockShard.SetEngine(h)
 	s.mockHistoryEngine = h
 
-	s.timerQueueActiveProcessor = newTimerQueueActiveProcessor(
+	s.timerQueueActiveTaskExecutor = newTimerQueueActiveTaskExecutor(
 		s.mockShard,
 		h,
-		s.mockMatchingClient,
-		newTaskAllocator(s.mockShard),
+		newTimerQueueActiveProcessor(
+			s.mockShard,
+			h,
+			s.mockMatchingClient,
+			newTaskAllocator(s.mockShard),
+			s.logger,
+		),
 		s.logger,
-	)
+		s.mockShard.GetMetricsClient(),
+		config,
+	).(*timerQueueActiveTaskExecutor)
 }
 
-func (s *timerQueueActiveProcessorSuite) TearDownTest() {
+func (s *timerQueueActiveTaskExecutorSuite) TearDownTest() {
 	s.controller.Finish()
 	s.mockShard.Finish(s.T())
 }
 
-func (s *timerQueueActiveProcessorSuite) TestProcessUserTimerTimeout_Fire() {
+func (s *timerQueueActiveTaskExecutorSuite) TestProcessUserTimerTimeout_Fire() {
 
 	execution := workflow.WorkflowExecution{
 		WorkflowId: common.StringPtr("some random workflow ID"),
@@ -220,14 +228,14 @@ func (s *timerQueueActiveProcessorSuite) TestProcessUserTimerTimeout_Fire() {
 	s.mockExecutionMgr.On("UpdateWorkflowExecution", mock.Anything).Return(&persistence.UpdateWorkflowExecutionResponse{MutableStateUpdateSessionStats: &persistence.MutableStateUpdateSessionStats{}}, nil).Once()
 
 	s.timeSource.Update(s.now.Add(2 * timerTimeout))
-	_, err = s.timerQueueActiveProcessor.process(newTaskInfo(nil, timerTask, s.logger))
+	err = s.timerQueueActiveTaskExecutor.execute(newTaskInfo(nil, timerTask, s.logger))
 	s.NoError(err)
 
 	_, ok := s.getMutableStateFromCache(s.domainID, execution.GetWorkflowId(), execution.GetRunId()).GetUserTimerInfo(timerID)
 	s.False(ok)
 }
 
-func (s *timerQueueActiveProcessorSuite) TestProcessUserTimerTimeout_Noop() {
+func (s *timerQueueActiveTaskExecutorSuite) TestProcessUserTimerTimeout_Noop() {
 
 	execution := workflow.WorkflowExecution{
 		WorkflowId: common.StringPtr("some random workflow ID"),
@@ -290,11 +298,11 @@ func (s *timerQueueActiveProcessorSuite) TestProcessUserTimerTimeout_Noop() {
 	s.mockExecutionMgr.On("GetWorkflowExecution", mock.Anything).Return(&persistence.GetWorkflowExecutionResponse{State: persistenceMutableState}, nil)
 
 	s.timeSource.Update(s.now.Add(2 * timerTimeout))
-	_, err = s.timerQueueActiveProcessor.process(newTaskInfo(nil, timerTask, s.logger))
+	err = s.timerQueueActiveTaskExecutor.execute(newTaskInfo(nil, timerTask, s.logger))
 	s.NoError(err)
 }
 
-func (s *timerQueueActiveProcessorSuite) TestProcessActivityTimeout_NoRetryPolicy_Fire() {
+func (s *timerQueueActiveTaskExecutorSuite) TestProcessActivityTimeout_NoRetryPolicy_Fire() {
 
 	execution := workflow.WorkflowExecution{
 		WorkflowId: common.StringPtr("some random workflow ID"),
@@ -363,14 +371,14 @@ func (s *timerQueueActiveProcessorSuite) TestProcessActivityTimeout_NoRetryPolic
 	s.mockExecutionMgr.On("UpdateWorkflowExecution", mock.Anything).Return(&persistence.UpdateWorkflowExecutionResponse{MutableStateUpdateSessionStats: &persistence.MutableStateUpdateSessionStats{}}, nil).Once()
 
 	s.timeSource.Update(s.now.Add(2 * timerTimeout))
-	_, err = s.timerQueueActiveProcessor.process(newTaskInfo(nil, timerTask, s.logger))
+	err = s.timerQueueActiveTaskExecutor.execute(newTaskInfo(nil, timerTask, s.logger))
 	s.NoError(err)
 
 	_, ok := s.getMutableStateFromCache(s.domainID, execution.GetWorkflowId(), execution.GetRunId()).GetActivityInfo(scheduledEvent.GetEventId())
 	s.False(ok)
 }
 
-func (s *timerQueueActiveProcessorSuite) TestProcessActivityTimeout_NoRetryPolicy_Noop() {
+func (s *timerQueueActiveTaskExecutorSuite) TestProcessActivityTimeout_NoRetryPolicy_Noop() {
 
 	execution := workflow.WorkflowExecution{
 		WorkflowId: common.StringPtr("some random workflow ID"),
@@ -441,11 +449,11 @@ func (s *timerQueueActiveProcessorSuite) TestProcessActivityTimeout_NoRetryPolic
 	s.mockExecutionMgr.On("GetWorkflowExecution", mock.Anything).Return(&persistence.GetWorkflowExecutionResponse{State: persistenceMutableState}, nil)
 
 	s.timeSource.Update(s.now.Add(2 * timerTimeout))
-	_, err = s.timerQueueActiveProcessor.process(newTaskInfo(nil, timerTask, s.logger))
+	err = s.timerQueueActiveTaskExecutor.execute(newTaskInfo(nil, timerTask, s.logger))
 	s.NoError(err)
 }
 
-func (s *timerQueueActiveProcessorSuite) TestProcessActivityTimeout_RetryPolicy_Retry() {
+func (s *timerQueueActiveTaskExecutorSuite) TestProcessActivityTimeout_RetryPolicy_Retry() {
 
 	execution := workflow.WorkflowExecution{
 		WorkflowId: common.StringPtr("some random workflow ID"),
@@ -525,7 +533,7 @@ func (s *timerQueueActiveProcessorSuite) TestProcessActivityTimeout_RetryPolicy_
 	s.mockExecutionMgr.On("UpdateWorkflowExecution", mock.Anything).Return(&persistence.UpdateWorkflowExecutionResponse{MutableStateUpdateSessionStats: &persistence.MutableStateUpdateSessionStats{}}, nil).Once()
 
 	s.timeSource.Update(s.now.Add(2 * timerTimeout))
-	_, err = s.timerQueueActiveProcessor.process(newTaskInfo(nil, timerTask, s.logger))
+	err = s.timerQueueActiveTaskExecutor.execute(newTaskInfo(nil, timerTask, s.logger))
 	s.NoError(err)
 
 	activityInfo, ok := s.getMutableStateFromCache(s.domainID, execution.GetWorkflowId(), execution.GetRunId()).GetActivityInfo(scheduledEvent.GetEventId())
@@ -536,7 +544,7 @@ func (s *timerQueueActiveProcessorSuite) TestProcessActivityTimeout_RetryPolicy_
 	s.Equal(int32(timerTaskStatusCreatedScheduleToStart), activityInfo.TimerTaskStatus)
 }
 
-func (s *timerQueueActiveProcessorSuite) TestProcessActivityTimeout_RetryPolicy_Fire() {
+func (s *timerQueueActiveTaskExecutorSuite) TestProcessActivityTimeout_RetryPolicy_Fire() {
 
 	execution := workflow.WorkflowExecution{
 		WorkflowId: common.StringPtr("some random workflow ID"),
@@ -614,14 +622,14 @@ func (s *timerQueueActiveProcessorSuite) TestProcessActivityTimeout_RetryPolicy_
 	s.mockExecutionMgr.On("UpdateWorkflowExecution", mock.Anything).Return(&persistence.UpdateWorkflowExecutionResponse{MutableStateUpdateSessionStats: &persistence.MutableStateUpdateSessionStats{}}, nil).Once()
 
 	s.timeSource.Update(s.now.Add(2 * timerTimeout))
-	_, err = s.timerQueueActiveProcessor.process(newTaskInfo(nil, timerTask, s.logger))
+	err = s.timerQueueActiveTaskExecutor.execute(newTaskInfo(nil, timerTask, s.logger))
 	s.NoError(err)
 
 	_, ok := s.getMutableStateFromCache(s.domainID, execution.GetWorkflowId(), execution.GetRunId()).GetActivityInfo(scheduledEvent.GetEventId())
 	s.False(ok)
 }
 
-func (s *timerQueueActiveProcessorSuite) TestProcessActivityTimeout_RetryPolicy_Noop() {
+func (s *timerQueueActiveTaskExecutorSuite) TestProcessActivityTimeout_RetryPolicy_Noop() {
 	execution := workflow.WorkflowExecution{
 		WorkflowId: common.StringPtr("some random workflow ID"),
 		RunId:      common.StringPtr(uuid.New()),
@@ -701,11 +709,11 @@ func (s *timerQueueActiveProcessorSuite) TestProcessActivityTimeout_RetryPolicy_
 	s.mockExecutionMgr.On("GetWorkflowExecution", mock.Anything).Return(&persistence.GetWorkflowExecutionResponse{State: persistenceMutableState}, nil)
 
 	s.timeSource.Update(s.now.Add(2 * timerTimeout))
-	_, err = s.timerQueueActiveProcessor.process(newTaskInfo(nil, timerTask, s.logger))
+	err = s.timerQueueActiveTaskExecutor.execute(newTaskInfo(nil, timerTask, s.logger))
 	s.NoError(err)
 }
 
-func (s *timerQueueActiveProcessorSuite) TestProcessActivityTimeout_Heartbeat_Noop() {
+func (s *timerQueueActiveTaskExecutorSuite) TestProcessActivityTimeout_Heartbeat_Noop() {
 	execution := workflow.WorkflowExecution{
 		WorkflowId: common.StringPtr("some random workflow ID"),
 		RunId:      common.StringPtr(uuid.New()),
@@ -784,11 +792,11 @@ func (s *timerQueueActiveProcessorSuite) TestProcessActivityTimeout_Heartbeat_No
 	persistenceMutableState := s.createPersistenceMutableState(mutableState, scheduledEvent.GetEventId(), scheduledEvent.GetVersion())
 	s.mockExecutionMgr.On("GetWorkflowExecution", mock.Anything).Return(&persistence.GetWorkflowExecutionResponse{State: persistenceMutableState}, nil)
 
-	_, err = s.timerQueueActiveProcessor.process(newTaskInfo(nil, timerTask, s.logger))
+	err = s.timerQueueActiveTaskExecutor.execute(newTaskInfo(nil, timerTask, s.logger))
 	s.NoError(err)
 }
 
-func (s *timerQueueActiveProcessorSuite) TestDecisionTimeout_Fire() {
+func (s *timerQueueActiveTaskExecutorSuite) TestDecisionTimeout_Fire() {
 
 	execution := workflow.WorkflowExecution{
 		WorkflowId: common.StringPtr("some random workflow ID"),
@@ -832,7 +840,7 @@ func (s *timerQueueActiveProcessorSuite) TestDecisionTimeout_Fire() {
 	s.mockHistoryV2Mgr.On("AppendHistoryNodes", mock.Anything).Return(&persistence.AppendHistoryNodesResponse{Size: 0}, nil).Once()
 	s.mockExecutionMgr.On("UpdateWorkflowExecution", mock.Anything).Return(&persistence.UpdateWorkflowExecutionResponse{MutableStateUpdateSessionStats: &persistence.MutableStateUpdateSessionStats{}}, nil).Once()
 
-	_, err = s.timerQueueActiveProcessor.process(newTaskInfo(nil, timerTask, s.logger))
+	err = s.timerQueueActiveTaskExecutor.execute(newTaskInfo(nil, timerTask, s.logger))
 	s.NoError(err)
 
 	decisionInfo, ok := s.getMutableStateFromCache(s.domainID, execution.GetWorkflowId(), execution.GetRunId()).GetPendingDecision()
@@ -842,7 +850,7 @@ func (s *timerQueueActiveProcessorSuite) TestDecisionTimeout_Fire() {
 	s.Equal(int64(1), decisionInfo.Attempt)
 }
 
-func (s *timerQueueActiveProcessorSuite) TestDecisionTimeout_Noop() {
+func (s *timerQueueActiveTaskExecutorSuite) TestDecisionTimeout_Noop() {
 
 	execution := workflow.WorkflowExecution{
 		WorkflowId: common.StringPtr("some random workflow ID"),
@@ -884,11 +892,11 @@ func (s *timerQueueActiveProcessorSuite) TestDecisionTimeout_Noop() {
 	persistenceMutableState := s.createPersistenceMutableState(mutableState, startedEvent.GetEventId(), startedEvent.GetVersion())
 	s.mockExecutionMgr.On("GetWorkflowExecution", mock.Anything).Return(&persistence.GetWorkflowExecutionResponse{State: persistenceMutableState}, nil).Once()
 
-	_, err = s.timerQueueActiveProcessor.process(newTaskInfo(nil, timerTask, s.logger))
+	err = s.timerQueueActiveTaskExecutor.execute(newTaskInfo(nil, timerTask, s.logger))
 	s.NoError(err)
 }
 
-func (s *timerQueueActiveProcessorSuite) TestWorkflowBackoffTimer_Fire() {
+func (s *timerQueueActiveTaskExecutorSuite) TestWorkflowBackoffTimer_Fire() {
 
 	execution := workflow.WorkflowExecution{
 		WorkflowId: common.StringPtr("some random workflow ID"),
@@ -929,7 +937,7 @@ func (s *timerQueueActiveProcessorSuite) TestWorkflowBackoffTimer_Fire() {
 	s.mockHistoryV2Mgr.On("AppendHistoryNodes", mock.Anything).Return(&persistence.AppendHistoryNodesResponse{Size: 0}, nil).Once()
 	s.mockExecutionMgr.On("UpdateWorkflowExecution", mock.Anything).Return(&persistence.UpdateWorkflowExecutionResponse{MutableStateUpdateSessionStats: &persistence.MutableStateUpdateSessionStats{}}, nil).Once()
 
-	_, err = s.timerQueueActiveProcessor.process(newTaskInfo(nil, timerTask, s.logger))
+	err = s.timerQueueActiveTaskExecutor.execute(newTaskInfo(nil, timerTask, s.logger))
 	s.NoError(err)
 
 	decisionInfo, ok := s.getMutableStateFromCache(s.domainID, execution.GetWorkflowId(), execution.GetRunId()).GetPendingDecision()
@@ -939,7 +947,7 @@ func (s *timerQueueActiveProcessorSuite) TestWorkflowBackoffTimer_Fire() {
 	s.Equal(int64(0), decisionInfo.Attempt)
 }
 
-func (s *timerQueueActiveProcessorSuite) TestWorkflowBackoffTimer_Noop() {
+func (s *timerQueueActiveTaskExecutorSuite) TestWorkflowBackoffTimer_Noop() {
 
 	execution := workflow.WorkflowExecution{
 		WorkflowId: common.StringPtr("some random workflow ID"),
@@ -983,11 +991,11 @@ func (s *timerQueueActiveProcessorSuite) TestWorkflowBackoffTimer_Noop() {
 	persistenceMutableState := s.createPersistenceMutableState(mutableState, event.GetEventId(), event.GetVersion())
 	s.mockExecutionMgr.On("GetWorkflowExecution", mock.Anything).Return(&persistence.GetWorkflowExecutionResponse{State: persistenceMutableState}, nil).Once()
 
-	_, err = s.timerQueueActiveProcessor.process(newTaskInfo(nil, timerTask, s.logger))
+	err = s.timerQueueActiveTaskExecutor.execute(newTaskInfo(nil, timerTask, s.logger))
 	s.NoError(err)
 }
 
-func (s *timerQueueActiveProcessorSuite) TestActivityRetryTimer_Fire() {
+func (s *timerQueueActiveTaskExecutorSuite) TestActivityRetryTimer_Fire() {
 
 	execution := workflow.WorkflowExecution{
 		WorkflowId: common.StringPtr("some random workflow ID"),
@@ -1071,11 +1079,11 @@ func (s *timerQueueActiveProcessorSuite) TestActivityRetryTimer_Fire() {
 		},
 	).Return(nil).Times(1)
 
-	_, err = s.timerQueueActiveProcessor.process(newTaskInfo(nil, timerTask, s.logger))
+	err = s.timerQueueActiveTaskExecutor.execute(newTaskInfo(nil, timerTask, s.logger))
 	s.NoError(err)
 }
 
-func (s *timerQueueActiveProcessorSuite) TestActivityRetryTimer_Noop() {
+func (s *timerQueueActiveTaskExecutorSuite) TestActivityRetryTimer_Noop() {
 
 	execution := workflow.WorkflowExecution{
 		WorkflowId: common.StringPtr("some random workflow ID"),
@@ -1148,11 +1156,11 @@ func (s *timerQueueActiveProcessorSuite) TestActivityRetryTimer_Noop() {
 	persistenceMutableState := s.createPersistenceMutableState(mutableState, scheduledEvent.GetEventId(), scheduledEvent.GetVersion())
 	s.mockExecutionMgr.On("GetWorkflowExecution", mock.Anything).Return(&persistence.GetWorkflowExecutionResponse{State: persistenceMutableState}, nil)
 
-	_, err = s.timerQueueActiveProcessor.process(newTaskInfo(nil, timerTask, s.logger))
+	err = s.timerQueueActiveTaskExecutor.execute(newTaskInfo(nil, timerTask, s.logger))
 	s.NoError(err)
 }
 
-func (s *timerQueueActiveProcessorSuite) TestWorkflowTimeout_Fire() {
+func (s *timerQueueActiveTaskExecutorSuite) TestWorkflowTimeout_Fire() {
 
 	execution := workflow.WorkflowExecution{
 		WorkflowId: common.StringPtr("some random workflow ID"),
@@ -1197,14 +1205,14 @@ func (s *timerQueueActiveProcessorSuite) TestWorkflowTimeout_Fire() {
 	s.mockHistoryV2Mgr.On("AppendHistoryNodes", mock.Anything).Return(&persistence.AppendHistoryNodesResponse{Size: 0}, nil).Once()
 	s.mockExecutionMgr.On("UpdateWorkflowExecution", mock.Anything).Return(&persistence.UpdateWorkflowExecutionResponse{MutableStateUpdateSessionStats: &persistence.MutableStateUpdateSessionStats{}}, nil).Once()
 
-	_, err = s.timerQueueActiveProcessor.process(newTaskInfo(nil, timerTask, s.logger))
+	err = s.timerQueueActiveTaskExecutor.execute(newTaskInfo(nil, timerTask, s.logger))
 	s.NoError(err)
 
 	running := s.getMutableStateFromCache(s.domainID, execution.GetWorkflowId(), execution.GetRunId()).IsWorkflowExecutionRunning()
 	s.False(running)
 }
 
-func (s *timerQueueActiveProcessorSuite) TestWorkflowTimeout_ContinueAsNew_Retry() {
+func (s *timerQueueActiveTaskExecutorSuite) TestWorkflowTimeout_ContinueAsNew_Retry() {
 
 	execution := workflow.WorkflowExecution{
 		WorkflowId: common.StringPtr("some random workflow ID"),
@@ -1258,7 +1266,7 @@ func (s *timerQueueActiveProcessorSuite) TestWorkflowTimeout_ContinueAsNew_Retry
 	s.mockHistoryV2Mgr.On("AppendHistoryNodes", mock.Anything).Return(&persistence.AppendHistoryNodesResponse{Size: 0}, nil).Times(2)
 	s.mockExecutionMgr.On("UpdateWorkflowExecution", mock.Anything).Return(&persistence.UpdateWorkflowExecutionResponse{MutableStateUpdateSessionStats: &persistence.MutableStateUpdateSessionStats{}}, nil).Once()
 
-	_, err = s.timerQueueActiveProcessor.process(newTaskInfo(nil, timerTask, s.logger))
+	err = s.timerQueueActiveTaskExecutor.execute(newTaskInfo(nil, timerTask, s.logger))
 	s.NoError(err)
 
 	state, closeStatus := s.getMutableStateFromCache(s.domainID, execution.GetWorkflowId(), execution.GetRunId()).GetWorkflowStateCloseStatus()
@@ -1266,7 +1274,7 @@ func (s *timerQueueActiveProcessorSuite) TestWorkflowTimeout_ContinueAsNew_Retry
 	s.Equal(persistence.WorkflowCloseStatusContinuedAsNew, closeStatus)
 }
 
-func (s *timerQueueActiveProcessorSuite) TestWorkflowTimeout_ContinueAsNew_Cron() {
+func (s *timerQueueActiveTaskExecutorSuite) TestWorkflowTimeout_ContinueAsNew_Cron() {
 
 	execution := workflow.WorkflowExecution{
 		WorkflowId: common.StringPtr("some random workflow ID"),
@@ -1315,7 +1323,7 @@ func (s *timerQueueActiveProcessorSuite) TestWorkflowTimeout_ContinueAsNew_Cron(
 	s.mockHistoryV2Mgr.On("AppendHistoryNodes", mock.Anything).Return(&persistence.AppendHistoryNodesResponse{Size: 0}, nil).Times(2)
 	s.mockExecutionMgr.On("UpdateWorkflowExecution", mock.Anything).Return(&persistence.UpdateWorkflowExecutionResponse{MutableStateUpdateSessionStats: &persistence.MutableStateUpdateSessionStats{}}, nil).Once()
 
-	_, err = s.timerQueueActiveProcessor.process(newTaskInfo(nil, timerTask, s.logger))
+	err = s.timerQueueActiveTaskExecutor.execute(newTaskInfo(nil, timerTask, s.logger))
 	s.NoError(err)
 
 	state, closeStatus := s.getMutableStateFromCache(s.domainID, execution.GetWorkflowId(), execution.GetRunId()).GetWorkflowStateCloseStatus()
@@ -1323,7 +1331,7 @@ func (s *timerQueueActiveProcessorSuite) TestWorkflowTimeout_ContinueAsNew_Cron(
 	s.Equal(persistence.WorkflowCloseStatusContinuedAsNew, closeStatus)
 }
 
-func (s *timerQueueActiveProcessorSuite) createPersistenceMutableState(
+func (s *timerQueueActiveTaskExecutorSuite) createPersistenceMutableState(
 	ms mutableState,
 	lastEventID int64,
 	lastEventVersion int64,
@@ -1343,7 +1351,7 @@ func (s *timerQueueActiveProcessorSuite) createPersistenceMutableState(
 	return createMutableState(ms)
 }
 
-func (s *timerQueueActiveProcessorSuite) getMutableStateFromCache(
+func (s *timerQueueActiveTaskExecutorSuite) getMutableStateFromCache(
 	domainID string,
 	workflowID string,
 	runID string,
