+++ service/history/timerQueueStandbyProcessor.go
@@ -21,13 +21,8 @@
 package history
 
 import (
-	"fmt"
 	"time"
 
-	workflow "github.com/uber/cadence/.gen/go/shared"
-	"github.com/uber/cadence/common"
-	"github.com/uber/cadence/common/clock"
-	"github.com/uber/cadence/common/cluster"
 	"github.com/uber/cadence/common/log"
 	"github.com/uber/cadence/common/log/tag"
 	"github.com/uber/cadence/common/metrics"
@@ -42,18 +37,12 @@ const (
 type (
 	timerQueueStandbyProcessorImpl struct {
 		shard                   ShardContext
-		config                  *Config
-		clusterMetadata         cluster.Metadata
-		historyService          *historyEngineImpl
-		cache                   *historyCache
 		timerTaskFilter         taskFilter
 		logger                  log.Logger
 		metricsClient           metrics.Client
-		clusterName             string
 		timerGate               RemoteTimerGate
 		timerQueueProcessorBase *timerQueueProcessorBase
-		historyRereplicator     xdc.HistoryRereplicator
-		nDCHistoryResender      xdc.NDCHistoryResender
+		taskExecutor            queueTaskExecutor
 	}
 )
 
@@ -96,14 +85,9 @@ func newTimerQueueStandbyProcessor(
 	)
 	processor := &timerQueueStandbyProcessorImpl{
 		shard:           shard,
-		config:          shard.GetConfig(),
-		clusterMetadata: shard.GetService().GetClusterMetadata(),
-		historyService:  historyService,
-		cache:           historyService.historyCache,
 		timerTaskFilter: timerTaskFilter,
 		logger:          logger,
 		metricsClient:   historyService.metricsClient,
-		clusterName:     clusterName,
 		timerGate:       timerGate,
 		timerQueueProcessorBase: newTimerQueueProcessorBase(
 			metrics.TimerStandbyQueueProcessorScope,
@@ -114,8 +98,16 @@ func newTimerQueueStandbyProcessor(
 			shard.GetConfig().TimerProcessorMaxPollRPS,
 			logger,
 		),
-		historyRereplicator: historyRereplicator,
-		nDCHistoryResender:  nDCHistoryResender,
+		taskExecutor: newTimerQueueStandbyTaskExecutor(
+			shard,
+			historyService,
+			historyRereplicator,
+			nDCHistoryResender,
+			logger,
+			historyService.metricsClient,
+			clusterName,
+			shard.GetConfig(),
+		),
 	}
 	processor.timerQueueProcessorBase.timerProcessor = processor
 	return processor
@@ -180,432 +172,440 @@ func (t *timerQueueStandbyProcessorImpl) complete(
 func (t *timerQueueStandbyProcessorImpl) process(
 	taskInfo *taskInfo,
 ) (int, error) {
-
-	timerTask, ok := taskInfo.task.(*persistence.TimerTaskInfo)
-	if !ok {
-		return metrics.TimerStandbyQueueProcessorScope, errUnexpectedQueueTask
-	}
-
-	var err error
-	switch timerTask.TaskType {
-	case persistence.TaskTypeUserTimer:
-		if taskInfo.shouldProcessTask {
-			err = t.processUserTimerTimeout(taskInfo)
-		}
-		return metrics.TimerStandbyTaskUserTimerScope, err
-
-	case persistence.TaskTypeActivityTimeout:
-		if taskInfo.shouldProcessTask {
-			err = t.processActivityTimeout(taskInfo)
-		}
-		return metrics.TimerStandbyTaskActivityTimeoutScope, err
-
-	case persistence.TaskTypeDecisionTimeout:
-		if taskInfo.shouldProcessTask {
-			err = t.processDecisionTimeout(taskInfo)
-		}
-		return metrics.TimerStandbyTaskDecisionTimeoutScope, err
-
-	case persistence.TaskTypeWorkflowTimeout:
-		// guarantee the processing of workflow execution history deletion
-		err = t.processWorkflowTimeout(taskInfo)
-		return metrics.TimerStandbyTaskWorkflowTimeoutScope, err
-
-	case persistence.TaskTypeActivityRetryTimer:
-		// retry backoff timer should not get created on passive cluster
-		return metrics.TimerStandbyTaskActivityRetryTimerScope, err
-
-	case persistence.TaskTypeWorkflowBackoffTimer:
-		if taskInfo.shouldProcessTask {
-			err = t.processWorkflowBackoffTimer(taskInfo)
-		}
-		return metrics.TimerStandbyTaskWorkflowBackoffTimerScope, err
-
-	case persistence.TaskTypeDeleteHistoryEvent:
-		// guarantee the processing of workflow execution history deletion
-		return metrics.TimerStandbyTaskDeleteHistoryEventScope, t.timerQueueProcessorBase.processDeleteHistoryEvent(timerTask)
-
-	default:
-		return metrics.TimerStandbyQueueProcessorScope, errUnknownTimerTask
-	}
+	// TODO: task metricScope should be determined when creating taskInfo
+	metricScope := t.timerQueueProcessorBase.getTimerTaskMetricScope(taskInfo.task.GetTaskType(), false)
+	return metricScope, t.taskExecutor.execute(taskInfo)
 }
 
-func (t *timerQueueStandbyProcessorImpl) processUserTimerTimeout(
-	taskInfo *taskInfo,
-) error {
+// func (t *timerQueueStandbyProcessorImpl) process(
+// 	taskInfo *taskInfo,
+// ) (int, error) {
 
-	actionFn := func(context workflowExecutionContext, mutableState mutableState) (interface{}, error) {
+// 	timerTask, ok := taskInfo.task.(*persistence.TimerTaskInfo)
+// 	if !ok {
+// 		return metrics.TimerStandbyQueueProcessorScope, errUnexpectedQueueTask
+// 	}
 
-		timerTask := taskInfo.task.(*persistence.TimerTaskInfo)
-		timerSequence := t.getTimerSequence(mutableState)
+// 	var err error
+// 	switch timerTask.TaskType {
+// 	case persistence.TaskTypeUserTimer:
+// 		if taskInfo.shouldProcessTask {
+// 			err = t.processUserTimerTimeout(taskInfo)
+// 		}
+// 		return metrics.TimerStandbyTaskUserTimerScope, err
 
-	Loop:
-		for _, timerSequenceID := range timerSequence.loadAndSortUserTimers() {
-			_, ok := mutableState.GetUserTimerInfoByEventID(timerSequenceID.eventID)
-			if !ok {
-				errString := fmt.Sprintf("failed to find in user timer event ID: %v", timerSequenceID.eventID)
-				t.logger.Error(errString)
-				return nil, &workflow.InternalServiceError{Message: errString}
-			}
+// 	case persistence.TaskTypeActivityTimeout:
+// 		if taskInfo.shouldProcessTask {
+// 			err = t.processActivityTimeout(taskInfo)
+// 		}
+// 		return metrics.TimerStandbyTaskActivityTimeoutScope, err
 
-			if isExpired := timerSequence.isExpired(
-				timerTask.VisibilityTimestamp,
-				timerSequenceID,
-			); isExpired {
-				return getHistoryResendInfo(mutableState)
-			}
-			// since the user timer are already sorted, so if there is one timer which will not expired
-			// all user timer after this timer will not expired
-			break Loop //nolint:staticcheck
-		}
-		// if there is no user timer expired, then we are good
-		return nil, nil
-	}
+// 	case persistence.TaskTypeDecisionTimeout:
+// 		if taskInfo.shouldProcessTask {
+// 			err = t.processDecisionTimeout(taskInfo)
+// 		}
+// 		return metrics.TimerStandbyTaskDecisionTimeoutScope, err
 
-	return t.processTimer(
-		taskInfo,
-		actionFn,
-		getStandbyPostActionFn(
-			taskInfo,
-			t.getCurrentTime,
-			t.config.StandbyTaskMissingEventsResendDelay(),
-			t.config.StandbyTaskMissingEventsDiscardDelay(),
-			t.fetchHistoryFromRemote,
-			standbyTimerTaskPostActionTaskDiscarded,
-		),
-	)
-}
+// 	case persistence.TaskTypeWorkflowTimeout:
+// 		// guarantee the processing of workflow execution history deletion
+// 		err = t.processWorkflowTimeout(taskInfo)
+// 		return metrics.TimerStandbyTaskWorkflowTimeoutScope, err
 
-func (t *timerQueueStandbyProcessorImpl) processActivityTimeout(
-	taskInfo *taskInfo,
-) error {
+// 	case persistence.TaskTypeActivityRetryTimer:
+// 		// retry backoff timer should not get created on passive cluster
+// 		return metrics.TimerStandbyTaskActivityRetryTimerScope, err
 
-	// activity heartbeat timer task is a special snowflake.
-	// normal activity timer task on the passive side will be generated by events related to activity in history replicator,
-	// and the standby timer processor will only need to verify whether the timer task can be safely throw away.
-	//
-	// activity heartbeat timer task cannot be handled in the way mentioned above.
-	// the reason is, there is no event driving the creation of new activity heartbeat timer.
-	// although there will be an task syncing activity from remote, the task is not an event,
-	// and cannot attempt to recreate a new activity timer task.
-	//
-	// the overall solution is to attempt to generate a new activity timer task whenever the
-	// task passed in is safe to be throw away.
+// 	case persistence.TaskTypeWorkflowBackoffTimer:
+// 		if taskInfo.shouldProcessTask {
+// 			err = t.processWorkflowBackoffTimer(taskInfo)
+// 		}
+// 		return metrics.TimerStandbyTaskWorkflowBackoffTimerScope, err
 
-	actionFn := func(context workflowExecutionContext, mutableState mutableState) (interface{}, error) {
+// 	case persistence.TaskTypeDeleteHistoryEvent:
+// 		// guarantee the processing of workflow execution history deletion
+// 		return metrics.TimerStandbyTaskDeleteHistoryEventScope, t.timerQueueProcessorBase.processDeleteHistoryEvent(timerTask)
 
-		timerTask := taskInfo.task.(*persistence.TimerTaskInfo)
-		timerSequence := t.getTimerSequence(mutableState)
-		updateMutableState := false
+// 	default:
+// 		return metrics.TimerStandbyQueueProcessorScope, errUnknownTimerTask
+// 	}
+// }
 
-	Loop:
-		for _, timerSequenceID := range timerSequence.loadAndSortActivityTimers() {
-			_, ok := mutableState.GetActivityInfo(timerSequenceID.eventID)
-			if !ok {
-				errString := fmt.Sprintf("failed to find in memory activity timer: %v", timerSequenceID.eventID)
-				t.logger.Error(errString)
-				return nil, &workflow.InternalServiceError{Message: errString}
-			}
+// func (t *timerQueueStandbyProcessorImpl) processUserTimerTimeout(
+// 	taskInfo *taskInfo,
+// ) error {
 
-			if isExpired := timerSequence.isExpired(
-				timerTask.VisibilityTimestamp,
-				timerSequenceID,
-			); isExpired {
-				return getHistoryResendInfo(mutableState)
-			}
-			// since the activity timer are already sorted, so if there is one timer which will not expired
-			// all activity timer after this timer will not expired
-			break Loop //nolint:staticcheck
-		}
+// 	actionFn := func(context workflowExecutionContext, mutableState mutableState) (interface{}, error) {
 
-		// for reason to update mutable state & generate a new activity task,
-		// see comments at the beginning of this function.
-		// NOTE: this is the only place in the standby logic where mutable state can be updated
+// 		timerTask := taskInfo.task.(*persistence.TimerTaskInfo)
+// 		timerSequence := t.getTimerSequence(mutableState)
 
-		// need to clear the activity heartbeat timer task marks
-		lastWriteVersion, err := mutableState.GetLastWriteVersion()
-		if err != nil {
-			return nil, err
-		}
-		isHeartBeatTask := timerTask.TimeoutType == int(workflow.TimeoutTypeHeartbeat)
-		if activityInfo, ok := mutableState.GetActivityInfo(timerTask.EventID); isHeartBeatTask && ok {
-			activityInfo.TimerTaskStatus = activityInfo.TimerTaskStatus &^ timerTaskStatusCreatedHeartbeat
-			if err := mutableState.UpdateActivity(activityInfo); err != nil {
-				return nil, err
-			}
-			updateMutableState = true
-		}
+// 	Loop:
+// 		for _, timerSequenceID := range timerSequence.loadAndSortUserTimers() {
+// 			_, ok := mutableState.GetUserTimerInfoByEventID(timerSequenceID.eventID)
+// 			if !ok {
+// 				errString := fmt.Sprintf("failed to find in user timer event ID: %v", timerSequenceID.eventID)
+// 				t.logger.Error(errString)
+// 				return nil, &workflow.InternalServiceError{Message: errString}
+// 			}
 
-		// passive logic need to explicitly call create timer
-		modified, err := timerSequence.createNextActivityTimer()
-		if err != nil {
-			return nil, err
-		}
-		updateMutableState = updateMutableState || modified
+// 			if isExpired := timerSequence.isExpired(
+// 				timerTask.VisibilityTimestamp,
+// 				timerSequenceID,
+// 			); isExpired {
+// 				return getHistoryResendInfo(mutableState)
+// 			}
+// 			// since the user timer are already sorted, so if there is one timer which will not expired
+// 			// all user timer after this timer will not expired
+// 			break Loop //nolint:staticcheck
+// 		}
+// 		// if there is no user timer expired, then we are good
+// 		return nil, nil
+// 	}
 
-		if !updateMutableState {
-			return nil, nil
-		}
+// 	return t.processTimer(
+// 		taskInfo,
+// 		actionFn,
+// 		getStandbyPostActionFn(
+// 			taskInfo,
+// 			t.getCurrentTime,
+// 			t.config.StandbyTaskMissingEventsResendDelay(),
+// 			t.config.StandbyTaskMissingEventsDiscardDelay(),
+// 			t.fetchHistoryFromRemote,
+// 			standbyTimerTaskPostActionTaskDiscarded,
+// 		),
+// 	)
+// }
 
-		now := t.getStandbyClusterTime()
-		// we need to handcraft some of the variables
-		// since the job being done here is update the activity and possibly write a timer task to DB
-		// also need to reset the current version.
-		if err := mutableState.UpdateCurrentVersion(lastWriteVersion, true); err != nil {
-			return nil, err
-		}
+// func (t *timerQueueStandbyProcessorImpl) processActivityTimeout(
+// 	taskInfo *taskInfo,
+// ) error {
 
-		err = context.updateWorkflowExecutionAsPassive(now)
-		return nil, err
-	}
+// 	// activity heartbeat timer task is a special snowflake.
+// 	// normal activity timer task on the passive side will be generated by events related to activity in history replicator,
+// 	// and the standby timer processor will only need to verify whether the timer task can be safely throw away.
+// 	//
+// 	// activity heartbeat timer task cannot be handled in the way mentioned above.
+// 	// the reason is, there is no event driving the creation of new activity heartbeat timer.
+// 	// although there will be an task syncing activity from remote, the task is not an event,
+// 	// and cannot attempt to recreate a new activity timer task.
+// 	//
+// 	// the overall solution is to attempt to generate a new activity timer task whenever the
+// 	// task passed in is safe to be throw away.
 
-	return t.processTimer(
-		taskInfo,
-		actionFn,
-		getStandbyPostActionFn(
-			taskInfo,
-			t.getCurrentTime,
-			t.config.StandbyTaskMissingEventsResendDelay(),
-			t.config.StandbyTaskMissingEventsDiscardDelay(),
-			t.fetchHistoryFromRemote,
-			standbyTimerTaskPostActionTaskDiscarded,
-		),
-	)
-}
+// 	actionFn := func(context workflowExecutionContext, mutableState mutableState) (interface{}, error) {
 
-func (t *timerQueueStandbyProcessorImpl) processDecisionTimeout(
-	taskInfo *taskInfo,
-) error {
+// 		timerTask := taskInfo.task.(*persistence.TimerTaskInfo)
+// 		timerSequence := t.getTimerSequence(mutableState)
+// 		updateMutableState := false
 
-	// decision schedule to start timer task is a special snowflake.
-	// the schedule to start timer is for sticky decision, which is
-	// not applicable on the passive cluster
-	if taskInfo.task.(*persistence.TimerTaskInfo).TimeoutType == int(workflow.TimeoutTypeScheduleToStart) {
-		return nil
-	}
+// 	Loop:
+// 		for _, timerSequenceID := range timerSequence.loadAndSortActivityTimers() {
+// 			_, ok := mutableState.GetActivityInfo(timerSequenceID.eventID)
+// 			if !ok {
+// 				errString := fmt.Sprintf("failed to find in memory activity timer: %v", timerSequenceID.eventID)
+// 				t.logger.Error(errString)
+// 				return nil, &workflow.InternalServiceError{Message: errString}
+// 			}
 
-	actionFn := func(context workflowExecutionContext, mutableState mutableState) (interface{}, error) {
+// 			if isExpired := timerSequence.isExpired(
+// 				timerTask.VisibilityTimestamp,
+// 				timerSequenceID,
+// 			); isExpired {
+// 				return getHistoryResendInfo(mutableState)
+// 			}
+// 			// since the activity timer are already sorted, so if there is one timer which will not expired
+// 			// all activity timer after this timer will not expired
+// 			break Loop //nolint:staticcheck
+// 		}
 
-		timerTask := taskInfo.task.(*persistence.TimerTaskInfo)
+// 		// for reason to update mutable state & generate a new activity task,
+// 		// see comments at the beginning of this function.
+// 		// NOTE: this is the only place in the standby logic where mutable state can be updated
 
-		decision, isPending := mutableState.GetDecisionInfo(timerTask.EventID)
-		if !isPending {
-			return nil, nil
-		}
+// 		// need to clear the activity heartbeat timer task marks
+// 		lastWriteVersion, err := mutableState.GetLastWriteVersion()
+// 		if err != nil {
+// 			return nil, err
+// 		}
+// 		isHeartBeatTask := timerTask.TimeoutType == int(workflow.TimeoutTypeHeartbeat)
+// 		if activityInfo, ok := mutableState.GetActivityInfo(timerTask.EventID); isHeartBeatTask && ok {
+// 			activityInfo.TimerTaskStatus = activityInfo.TimerTaskStatus &^ timerTaskStatusCreatedHeartbeat
+// 			if err := mutableState.UpdateActivity(activityInfo); err != nil {
+// 				return nil, err
+// 			}
+// 			updateMutableState = true
+// 		}
 
-		ok, err := verifyTaskVersion(t.shard, t.logger, timerTask.DomainID, decision.Version, timerTask.Version, timerTask)
-		if err != nil || !ok {
-			return nil, err
-		}
+// 		// passive logic need to explicitly call create timer
+// 		modified, err := timerSequence.createNextActivityTimer()
+// 		if err != nil {
+// 			return nil, err
+// 		}
+// 		updateMutableState = updateMutableState || modified
 
-		return getHistoryResendInfo(mutableState)
-	}
+// 		if !updateMutableState {
+// 			return nil, nil
+// 		}
 
-	return t.processTimer(
-		taskInfo,
-		actionFn,
-		getStandbyPostActionFn(
-			taskInfo,
-			t.getCurrentTime,
-			t.config.StandbyTaskMissingEventsResendDelay(),
-			t.config.StandbyTaskMissingEventsDiscardDelay(),
-			t.fetchHistoryFromRemote,
-			standbyTimerTaskPostActionTaskDiscarded,
-		),
-	)
-}
+// 		now := t.getStandbyClusterTime()
+// 		// we need to handcraft some of the variables
+// 		// since the job being done here is update the activity and possibly write a timer task to DB
+// 		// also need to reset the current version.
+// 		if err := mutableState.UpdateCurrentVersion(lastWriteVersion, true); err != nil {
+// 			return nil, err
+// 		}
 
-func (t *timerQueueStandbyProcessorImpl) processWorkflowBackoffTimer(
-	taskInfo *taskInfo,
-) error {
+// 		err = context.updateWorkflowExecutionAsPassive(now)
+// 		return nil, err
+// 	}
 
-	actionFn := func(context workflowExecutionContext, mutableState mutableState) (interface{}, error) {
+// 	return t.processTimer(
+// 		taskInfo,
+// 		actionFn,
+// 		getStandbyPostActionFn(
+// 			taskInfo,
+// 			t.getCurrentTime,
+// 			t.config.StandbyTaskMissingEventsResendDelay(),
+// 			t.config.StandbyTaskMissingEventsDiscardDelay(),
+// 			t.fetchHistoryFromRemote,
+// 			standbyTimerTaskPostActionTaskDiscarded,
+// 		),
+// 	)
+// }
 
-		if mutableState.HasProcessedOrPendingDecision() {
-			// if there is one decision already been processed
-			// or has pending decision, meaning workflow has already running
-			return nil, nil
-		}
+// func (t *timerQueueStandbyProcessorImpl) processDecisionTimeout(
+// 	taskInfo *taskInfo,
+// ) error {
 
-		// Note: do not need to verify task version here
-		// logic can only go here if mutable state build's next event ID is 2
-		// meaning history only contains workflow started event.
-		// we can do the checking of task version vs workflow started version
-		// however, workflow started version is immutable
+// 	// decision schedule to start timer task is a special snowflake.
+// 	// the schedule to start timer is for sticky decision, which is
+// 	// not applicable on the passive cluster
+// 	if taskInfo.task.(*persistence.TimerTaskInfo).TimeoutType == int(workflow.TimeoutTypeScheduleToStart) {
+// 		return nil
+// 	}
 
-		// active cluster will add first decision task after backoff timeout.
-		// standby cluster should just call ack manager to retry this task
-		// since we are stilling waiting for the first DecisionScheduledEvent to be replicated from active side.
+// 	actionFn := func(context workflowExecutionContext, mutableState mutableState) (interface{}, error) {
 
-		return getHistoryResendInfo(mutableState)
-	}
+// 		timerTask := taskInfo.task.(*persistence.TimerTaskInfo)
 
-	return t.processTimer(
-		taskInfo,
-		actionFn,
-		getStandbyPostActionFn(
-			taskInfo,
-			t.getCurrentTime,
-			t.config.StandbyTaskMissingEventsResendDelay(),
-			t.config.StandbyTaskMissingEventsDiscardDelay(),
-			t.fetchHistoryFromRemote,
-			standbyTimerTaskPostActionTaskDiscarded,
-		),
-	)
-}
+// 		decision, isPending := mutableState.GetDecisionInfo(timerTask.EventID)
+// 		if !isPending {
+// 			return nil, nil
+// 		}
 
-func (t *timerQueueStandbyProcessorImpl) processWorkflowTimeout(
-	taskInfo *taskInfo,
-) error {
+// 		ok, err := verifyTaskVersion(t.shard, t.logger, timerTask.DomainID, decision.Version, timerTask.Version, timerTask)
+// 		if err != nil || !ok {
+// 			return nil, err
+// 		}
 
-	actionFn := func(context workflowExecutionContext, mutableState mutableState) (interface{}, error) {
+// 		return getHistoryResendInfo(mutableState)
+// 	}
 
-		timerTask := taskInfo.task.(*persistence.TimerTaskInfo)
+// 	return t.processTimer(
+// 		taskInfo,
+// 		actionFn,
+// 		getStandbyPostActionFn(
+// 			taskInfo,
+// 			t.getCurrentTime,
+// 			t.config.StandbyTaskMissingEventsResendDelay(),
+// 			t.config.StandbyTaskMissingEventsDiscardDelay(),
+// 			t.fetchHistoryFromRemote,
+// 			standbyTimerTaskPostActionTaskDiscarded,
+// 		),
+// 	)
+// }
 
-		// we do not need to notify new timer to base, since if there is no new event being replicated
-		// checking again if the timer can be completed is meaningless
+// func (t *timerQueueStandbyProcessorImpl) processWorkflowBackoffTimer(
+// 	taskInfo *taskInfo,
+// ) error {
 
-		startVersion, err := mutableState.GetStartVersion()
-		if err != nil {
-			return nil, err
-		}
-		ok, err := verifyTaskVersion(t.shard, t.logger, timerTask.DomainID, startVersion, timerTask.Version, timerTask)
-		if err != nil || !ok {
-			return nil, err
-		}
+// 	actionFn := func(context workflowExecutionContext, mutableState mutableState) (interface{}, error) {
 
-		return getHistoryResendInfo(mutableState)
-	}
+// 		if mutableState.HasProcessedOrPendingDecision() {
+// 			// if there is one decision already been processed
+// 			// or has pending decision, meaning workflow has already running
+// 			return nil, nil
+// 		}
 
-	return t.processTimer(
-		taskInfo,
-		actionFn,
-		getStandbyPostActionFn(
-			taskInfo,
-			t.getCurrentTime,
-			t.config.StandbyTaskMissingEventsResendDelay(),
-			t.config.StandbyTaskMissingEventsDiscardDelay(),
-			t.fetchHistoryFromRemote,
-			standbyTimerTaskPostActionTaskDiscarded,
-		),
-	)
-}
+// 		// Note: do not need to verify task version here
+// 		// logic can only go here if mutable state build's next event ID is 2
+// 		// meaning history only contains workflow started event.
+// 		// we can do the checking of task version vs workflow started version
+// 		// however, workflow started version is immutable
 
-func (t *timerQueueStandbyProcessorImpl) getStandbyClusterTime() time.Time {
-	// time of remote cluster in the shard is delayed by "StandbyClusterDelay"
-	// so to get the current accurate remote cluster time, need to add it back
-	return t.shard.GetCurrentTime(t.clusterName).Add(t.shard.GetConfig().StandbyClusterDelay())
-}
+// 		// active cluster will add first decision task after backoff timeout.
+// 		// standby cluster should just call ack manager to retry this task
+// 		// since we are stilling waiting for the first DecisionScheduledEvent to be replicated from active side.
 
-func (t *timerQueueStandbyProcessorImpl) getTimerSequence(
-	mutableState mutableState,
-) timerSequence {
+// 		return getHistoryResendInfo(mutableState)
+// 	}
 
-	timeSource := clock.NewEventTimeSource()
-	now := t.getStandbyClusterTime()
-	timeSource.Update(now)
-	return newTimerSequence(timeSource, mutableState)
-}
+// 	return t.processTimer(
+// 		taskInfo,
+// 		actionFn,
+// 		getStandbyPostActionFn(
+// 			taskInfo,
+// 			t.getCurrentTime,
+// 			t.config.StandbyTaskMissingEventsResendDelay(),
+// 			t.config.StandbyTaskMissingEventsDiscardDelay(),
+// 			t.fetchHistoryFromRemote,
+// 			standbyTimerTaskPostActionTaskDiscarded,
+// 		),
+// 	)
+// }
 
-func (t *timerQueueStandbyProcessorImpl) processTimer(
-	taskInfo *taskInfo,
-	actionFn standbyActionFn,
-	postActionFn standbyPostActionFn,
-) (retError error) {
+// func (t *timerQueueStandbyProcessorImpl) processWorkflowTimeout(
+// 	taskInfo *taskInfo,
+// ) error {
 
-	timerTask := taskInfo.task.(*persistence.TimerTaskInfo)
-	context, release, err := t.cache.getOrCreateWorkflowExecutionForBackground(
-		t.timerQueueProcessorBase.getDomainIDAndWorkflowExecution(timerTask),
-	)
-	if err != nil {
-		return err
-	}
-	defer func() {
-		if retError == ErrTaskRetry {
-			release(nil)
-		} else {
-			release(retError)
-		}
-	}()
+// 	actionFn := func(context workflowExecutionContext, mutableState mutableState) (interface{}, error) {
 
-	mutableState, err := loadMutableStateForTimerTask(context, timerTask, t.metricsClient, t.logger)
-	if err != nil {
-		return err
-	}
-	if mutableState == nil {
-		return nil
-	}
+// 		timerTask := taskInfo.task.(*persistence.TimerTaskInfo)
 
-	if !mutableState.IsWorkflowExecutionRunning() {
-		// workflow already finished, no need to process the timer
-		return nil
-	}
+// 		// we do not need to notify new timer to base, since if there is no new event being replicated
+// 		// checking again if the timer can be completed is meaningless
 
-	historyResendInfo, err := actionFn(context, mutableState)
-	if err != nil {
-		return err
-	}
+// 		startVersion, err := mutableState.GetStartVersion()
+// 		if err != nil {
+// 			return nil, err
+// 		}
+// 		ok, err := verifyTaskVersion(t.shard, t.logger, timerTask.DomainID, startVersion, timerTask.Version, timerTask)
+// 		if err != nil || !ok {
+// 			return nil, err
+// 		}
 
-	release(nil)
-	return postActionFn(taskInfo, historyResendInfo, taskInfo.logger)
-}
+// 		return getHistoryResendInfo(mutableState)
+// 	}
 
-func (t *timerQueueStandbyProcessorImpl) fetchHistoryFromRemote(
-	taskInfo *taskInfo,
-	postActionInfo interface{},
-	log log.Logger,
-) error {
+// 	return t.processTimer(
+// 		taskInfo,
+// 		actionFn,
+// 		getStandbyPostActionFn(
+// 			taskInfo,
+// 			t.getCurrentTime,
+// 			t.config.StandbyTaskMissingEventsResendDelay(),
+// 			t.config.StandbyTaskMissingEventsDiscardDelay(),
+// 			t.fetchHistoryFromRemote,
+// 			standbyTimerTaskPostActionTaskDiscarded,
+// 		),
+// 	)
+// }
 
-	if postActionInfo == nil {
-		return nil
-	}
+// func (t *timerQueueStandbyProcessorImpl) getStandbyClusterTime() time.Time {
+// 	// time of remote cluster in the shard is delayed by "StandbyClusterDelay"
+// 	// so to get the current accurate remote cluster time, need to add it back
+// 	return t.shard.GetCurrentTime(t.clusterName).Add(t.shard.GetConfig().StandbyClusterDelay())
+// }
 
-	timerTask := taskInfo.task.(*persistence.TimerTaskInfo)
-	resendInfo := postActionInfo.(*historyResendInfo)
+// func (t *timerQueueStandbyProcessorImpl) getTimerSequence(
+// 	mutableState mutableState,
+// ) timerSequence {
 
-	t.metricsClient.IncCounter(metrics.HistoryRereplicationByTimerTaskScope, metrics.CadenceClientRequests)
-	stopwatch := t.metricsClient.StartTimer(metrics.HistoryRereplicationByTimerTaskScope, metrics.CadenceClientLatency)
-	defer stopwatch.Stop()
+// 	timeSource := clock.NewEventTimeSource()
+// 	now := t.getStandbyClusterTime()
+// 	timeSource.Update(now)
+// 	return newTimerSequence(timeSource, mutableState)
+// }
 
-	var err error
-	if resendInfo.lastEventID != nil && resendInfo.lastEventVersion != nil {
-		err = t.nDCHistoryResender.SendSingleWorkflowHistory(
-			timerTask.DomainID,
-			timerTask.WorkflowID,
-			timerTask.RunID,
-			resendInfo.lastEventID,
-			resendInfo.lastEventVersion,
-			nil,
-			nil,
-		)
-	} else if resendInfo.nextEventID != nil {
-		err = t.historyRereplicator.SendMultiWorkflowHistory(
-			timerTask.DomainID,
-			timerTask.WorkflowID,
-			timerTask.RunID,
-			*resendInfo.nextEventID,
-			timerTask.RunID,
-			common.EndEventID, // use common.EndEventID since we do not know where is the end
-		)
-	} else {
-		err = &workflow.InternalServiceError{
-			Message: "timerQueueStandbyProcessor encounter empty historyResendInfo",
-		}
-	}
+// func (t *timerQueueStandbyProcessorImpl) processTimer(
+// 	taskInfo *taskInfo,
+// 	actionFn standbyActionFn,
+// 	postActionFn standbyPostActionFn,
+// ) (retError error) {
 
-	if err != nil {
-		t.logger.Error("Error re-replicating history from remote.",
-			tag.ShardID(t.shard.GetShardID()),
-			tag.WorkflowDomainID(timerTask.DomainID),
-			tag.WorkflowID(timerTask.WorkflowID),
-			tag.WorkflowRunID(timerTask.RunID),
-			tag.ClusterName(t.clusterName))
-	}
+// 	timerTask := taskInfo.task.(*persistence.TimerTaskInfo)
+// 	context, release, err := t.cache.getOrCreateWorkflowExecutionForBackground(
+// 		t.timerQueueProcessorBase.getDomainIDAndWorkflowExecution(timerTask),
+// 	)
+// 	if err != nil {
+// 		return err
+// 	}
+// 	defer func() {
+// 		if retError == ErrTaskRetry {
+// 			release(nil)
+// 		} else {
+// 			release(retError)
+// 		}
+// 	}()
 
-	// return error so task processing logic will retry
-	return ErrTaskRetry
-}
+// 	mutableState, err := loadMutableStateForTimerTask(context, timerTask, t.metricsClient, t.logger)
+// 	if err != nil {
+// 		return err
+// 	}
+// 	if mutableState == nil {
+// 		return nil
+// 	}
 
-func (t *timerQueueStandbyProcessorImpl) getCurrentTime() time.Time {
-	return t.shard.GetCurrentTime(t.clusterName)
-}
+// 	if !mutableState.IsWorkflowExecutionRunning() {
+// 		// workflow already finished, no need to process the timer
+// 		return nil
+// 	}
+
+// 	historyResendInfo, err := actionFn(context, mutableState)
+// 	if err != nil {
+// 		return err
+// 	}
+
+// 	release(nil)
+// 	return postActionFn(taskInfo, historyResendInfo, taskInfo.logger)
+// }
+
+// func (t *timerQueueStandbyProcessorImpl) fetchHistoryFromRemote(
+// 	taskInfo *taskInfo,
+// 	postActionInfo interface{},
+// 	log log.Logger,
+// ) error {
+
+// 	if postActionInfo == nil {
+// 		return nil
+// 	}
+
+// 	timerTask := taskInfo.task.(*persistence.TimerTaskInfo)
+// 	resendInfo := postActionInfo.(*historyResendInfo)
+
+// 	t.metricsClient.IncCounter(metrics.HistoryRereplicationByTimerTaskScope, metrics.CadenceClientRequests)
+// 	stopwatch := t.metricsClient.StartTimer(metrics.HistoryRereplicationByTimerTaskScope, metrics.CadenceClientLatency)
+// 	defer stopwatch.Stop()
+
+// 	var err error
+// 	if resendInfo.lastEventID != nil && resendInfo.lastEventVersion != nil {
+// 		err = t.nDCHistoryResender.SendSingleWorkflowHistory(
+// 			timerTask.DomainID,
+// 			timerTask.WorkflowID,
+// 			timerTask.RunID,
+// 			resendInfo.lastEventID,
+// 			resendInfo.lastEventVersion,
+// 			nil,
+// 			nil,
+// 		)
+// 	} else if resendInfo.nextEventID != nil {
+// 		err = t.historyRereplicator.SendMultiWorkflowHistory(
+// 			timerTask.DomainID,
+// 			timerTask.WorkflowID,
+// 			timerTask.RunID,
+// 			*resendInfo.nextEventID,
+// 			timerTask.RunID,
+// 			common.EndEventID, // use common.EndEventID since we do not know where is the end
+// 		)
+// 	} else {
+// 		err = &workflow.InternalServiceError{
+// 			Message: "timerQueueStandbyProcessor encounter empty historyResendInfo",
+// 		}
+// 	}
+
+// 	if err != nil {
+// 		t.logger.Error("Error re-replicating history from remote.",
+// 			tag.ShardID(t.shard.GetShardID()),
+// 			tag.WorkflowDomainID(timerTask.DomainID),
+// 			tag.WorkflowID(timerTask.WorkflowID),
+// 			tag.WorkflowRunID(timerTask.RunID),
+// 			tag.ClusterName(t.clusterName))
+// 	}
+
+// 	// return error so task processing logic will retry
+// 	return ErrTaskRetry
+// }
+
+// func (t *timerQueueStandbyProcessorImpl) getCurrentTime() time.Time {
+// 	return t.shard.GetCurrentTime(t.clusterName)
+// }
